@@ -5,32 +5,32 @@ import (
 )
 
 // ConfigOption is a marker interface for options to NewConfiguration.
-type ConfigOption interface{}
+type ConfigOption any
 
 // NodeListOption must be implemented by node providers.
 type NodeListOption interface {
 	ConfigOption
-	newConfig(*RawManager) (RawConfiguration, error)
+	newConfig(*manager) (RawConfiguration, error)
 }
 
 type nodeIDMap struct {
 	idMap map[string]uint32
 }
 
-func (o nodeIDMap) newConfig(mgr *RawManager) (nodes RawConfiguration, err error) {
+func (o nodeIDMap) newConfig(mgr *manager) (cfg RawConfiguration, err error) {
 	if len(o.idMap) == 0 {
-		return nil, fmt.Errorf("config: missing required node map")
+		return RawConfiguration{}, fmt.Errorf("config: missing required node map")
 	}
-	nodes = make(RawConfiguration, 0, len(o.idMap))
+	nodes := make([]*RawNode, 0, len(o.idMap))
 	for naddr, id := range o.idMap {
-		node, found := mgr.Node(id)
+		node, found := mgr.node(id)
 		if !found {
 			node, err = NewRawNodeWithID(naddr, id)
 			if err != nil {
-				return nil, err
+				return RawConfiguration{}, err
 			}
-			if err = mgr.AddNode(node); err != nil {
-				return nil, err
+			if err = mgr.addNode(node); err != nil {
+				return RawConfiguration{}, err
 			}
 		}
 		nodes = append(nodes, node)
@@ -38,7 +38,13 @@ func (o nodeIDMap) newConfig(mgr *RawManager) (nodes RawConfiguration, err error
 	// Sort nodes to ensure deterministic iteration.
 	OrderedBy(ID).Sort(mgr.nodes)
 	OrderedBy(ID).Sort(nodes)
-	return nodes, nil
+
+	cfg = RawConfiguration{
+		RawNodes: nodes,
+		mgr:      mgr,
+	}
+
+	return cfg, nil
 }
 
 // WithNodeMap returns a NodeListOption containing the provided
@@ -51,19 +57,19 @@ type nodeList struct {
 	addrsList []string
 }
 
-func (o nodeList) newConfig(mgr *RawManager) (nodes RawConfiguration, err error) {
+func (o nodeList) newConfig(mgr *manager) (cfg RawConfiguration, err error) {
 	if len(o.addrsList) == 0 {
-		return nil, fmt.Errorf("config: missing required node addresses")
+		return RawConfiguration{}, fmt.Errorf("config: missing required node addresses")
 	}
-	nodes = make(RawConfiguration, 0, len(o.addrsList))
+	nodes := make([]*RawNode, 0, len(o.addrsList))
 	for _, naddr := range o.addrsList {
 		node, err := NewRawNode(naddr)
 		if err != nil {
-			return nil, err
+			return RawConfiguration{}, err
 		}
-		if n, found := mgr.Node(node.ID()); !found {
-			if err = mgr.AddNode(node); err != nil {
-				return nil, err
+		if n, found := mgr.node(node.ID()); !found {
+			if err = mgr.addNode(node); err != nil {
+				return RawConfiguration{}, err
 			}
 		} else {
 			node = n
@@ -73,11 +79,17 @@ func (o nodeList) newConfig(mgr *RawManager) (nodes RawConfiguration, err error)
 	// Sort nodes to ensure deterministic iteration.
 	OrderedBy(ID).Sort(mgr.nodes)
 	OrderedBy(ID).Sort(nodes)
-	return nodes, nil
+
+	cfg = RawConfiguration{
+		RawNodes: nodes,
+		mgr:      mgr,
+	}
+
+	return cfg, nil
 }
 
 // WithNodeList returns a NodeListOption containing the provided list of node addresses.
-// With this option, node IDs are generated by the Manager.
+// With this option, node IDs are generated by the manager.
 func WithNodeList(addrsList []string) NodeListOption {
 	return &nodeList{addrsList: addrsList}
 }
@@ -86,23 +98,29 @@ type nodeIDs struct {
 	nodeIDs []uint32
 }
 
-func (o nodeIDs) newConfig(mgr *RawManager) (nodes RawConfiguration, err error) {
+func (o nodeIDs) newConfig(mgr *manager) (cfg RawConfiguration, err error) {
 	if len(o.nodeIDs) == 0 {
-		return nil, fmt.Errorf("config: missing required node IDs")
+		return RawConfiguration{}, fmt.Errorf("config: missing required node IDs")
 	}
-	nodes = make(RawConfiguration, 0, len(o.nodeIDs))
+	nodes := make([]*RawNode, 0, len(o.nodeIDs))
 	for _, id := range o.nodeIDs {
-		node, found := mgr.Node(id)
+		node, found := mgr.node(id)
 		if !found {
 			// Node IDs must have been registered previously
-			return nil, fmt.Errorf("config: node %d not found", id)
+			return RawConfiguration{}, fmt.Errorf("config: node %d not found", id)
 		}
 		nodes = append(nodes, node)
 	}
 	// Sort nodes to ensure deterministic iteration.
 	OrderedBy(ID).Sort(mgr.nodes)
 	OrderedBy(ID).Sort(nodes)
-	return nodes, nil
+
+	cfg = RawConfiguration{
+		RawNodes: nodes,
+		mgr:      mgr,
+	}
+
+	return cfg, nil
 }
 
 // WithNodeIDs returns a NodeListOption containing a list of node IDs.
@@ -116,10 +134,10 @@ type addNodes struct {
 	new NodeListOption
 }
 
-func (o addNodes) newConfig(mgr *RawManager) (nodes RawConfiguration, err error) {
+func (o addNodes) newConfig(mgr *manager) (nodes RawConfiguration, err error) {
 	newNodes, err := o.new.newConfig(mgr)
 	if err != nil {
-		return nil, err
+		return RawConfiguration{}, err
 	}
 	ac := &addConfig{old: o.old, add: newNodes}
 	return ac.newConfig(mgr)
@@ -136,10 +154,10 @@ type addConfig struct {
 	add RawConfiguration
 }
 
-func (o addConfig) newConfig(mgr *RawManager) (nodes RawConfiguration, err error) {
-	nodes = make(RawConfiguration, 0, len(o.old)+len(o.add))
+func (o addConfig) newConfig(mgr *manager) (cfg RawConfiguration, err error) {
+	nodes := make([]*RawNode, 0, len(o.old.RawNodes)+len(o.add.RawNodes))
 	m := make(map[uint32]bool)
-	for _, n := range append(o.old, o.add...) {
+	for _, n := range append(o.old.RawNodes, o.add.RawNodes...) {
 		if !m[n.id] {
 			m[n.id] = true
 			nodes = append(nodes, n)
@@ -148,7 +166,13 @@ func (o addConfig) newConfig(mgr *RawManager) (nodes RawConfiguration, err error
 	// Sort nodes to ensure deterministic iteration.
 	OrderedBy(ID).Sort(mgr.nodes)
 	OrderedBy(ID).Sort(nodes)
-	return nodes, err
+
+	cfg = RawConfiguration{
+		RawNodes: nodes,
+		mgr:      mgr,
+	}
+
+	return cfg, err
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
@@ -163,8 +187,8 @@ func (c RawConfiguration) WithoutNodes(ids ...uint32) NodeListOption {
 	for _, id := range ids {
 		rmIDs[id] = true
 	}
-	keepIDs := make([]uint32, 0, len(c))
-	for _, cNode := range c {
+	keepIDs := make([]uint32, 0, len(c.RawNodes))
+	for _, cNode := range c.RawNodes {
 		if !rmIDs[cNode.id] {
 			keepIDs = append(keepIDs, cNode.id)
 		}
@@ -176,11 +200,11 @@ func (c RawConfiguration) WithoutNodes(ids ...uint32) NodeListOption {
 // from c without the nodes in rm.
 func (c RawConfiguration) Except(rm RawConfiguration) NodeListOption {
 	rmIDs := make(map[uint32]bool)
-	for _, rmNode := range rm {
+	for _, rmNode := range rm.RawNodes {
 		rmIDs[rmNode.id] = true
 	}
-	keepIDs := make([]uint32, 0, len(c))
-	for _, cNode := range c {
+	keepIDs := make([]uint32, 0, len(c.RawNodes))
+	for _, cNode := range c.RawNodes {
 		if !rmIDs[cNode.id] {
 			keepIDs = append(keepIDs, cNode.id)
 		}
