@@ -18,7 +18,7 @@ type onewaySrv struct {
 	received  chan *oneway.Request
 }
 
-func (s *onewaySrv) Unicast(ctx gorums.ServerCtx, r *oneway.Request) {
+func (s *onewaySrv) Unicast(_ gorums.ServerCtx, r *oneway.Request) {
 	if s.benchmark {
 		return
 	}
@@ -26,7 +26,7 @@ func (s *onewaySrv) Unicast(ctx gorums.ServerCtx, r *oneway.Request) {
 	s.wg.Done()
 }
 
-func (s *onewaySrv) Multicast(ctx gorums.ServerCtx, r *oneway.Request) {
+func (s *onewaySrv) Multicast(_ gorums.ServerCtx, r *oneway.Request) {
 	if s.benchmark {
 		return
 	}
@@ -34,20 +34,18 @@ func (s *onewaySrv) Multicast(ctx gorums.ServerCtx, r *oneway.Request) {
 	s.wg.Done()
 }
 
-func (s *onewaySrv) MulticastPerNode(ctx gorums.ServerCtx, r *oneway.Request) {
+func (s *onewaySrv) MulticastPerNode(_ gorums.ServerCtx, r *oneway.Request) {
 	if s.benchmark {
 		return
 	}
 	s.received <- r
 	s.wg.Done()
 }
-
-type testQSpec struct{}
 
 func setup(t testing.TB, cfgSize int) (cfg *oneway.OnewayTestConfiguration, srvs []*onewaySrv, teardown func()) {
 	t.Helper()
 	srvs = make([]*onewaySrv, cfgSize)
-	for i := 0; i < cfgSize; i++ {
+	for i := range cfgSize {
 		srvs[i] = &onewaySrv{received: make(chan *oneway.Request, numCalls)}
 	}
 
@@ -61,17 +59,17 @@ func setup(t testing.TB, cfgSize int) (cfg *oneway.OnewayTestConfiguration, srvs
 		nodeMap[addr] = uint32(i)
 	}
 
-	mgr := oneway.NewOnewayTestManager(
+	cfg, err := oneway.NewOnewayTestConfiguration(
+		gorums.WithNodeMap(nodeMap),
 		gorums.WithGrpcDialOptions(
 			grpc.WithTransportCredentials(insecure.NewCredentials()),
 		),
 	)
-	cfg, err := mgr.NewConfiguration(gorums.WithNodeMap(nodeMap))
 	if err != nil {
 		t.Fatal(err)
 	}
 	teardown = func() {
-		mgr.Close()
+		cfg.Close()
 		closeServers()
 	}
 	return cfg, srvs, teardown
@@ -106,12 +104,12 @@ func TestOnewayCalls(t *testing.T) {
 	for _, test := range tests {
 		t.Run(fmt.Sprintf("%s/Servers=%d", test.name, test.servers), func(t *testing.T) {
 			cfg, srvs, teardown := setup(t, test.servers)
-			for i := 0; i < len(srvs); i++ {
+			for i := range srvs {
 				srvs[i].wg.Add(test.calls)
 			}
 
 			for c := 1; c <= test.calls; c++ {
-				in := &oneway.Request{Num: uint64(c)}
+				in := oneway.Request_builder{Num: uint64(c)}.Build()
 				if test.sendWait {
 					f(cfg)(context.Background(), in)
 				} else {
@@ -120,13 +118,13 @@ func TestOnewayCalls(t *testing.T) {
 			}
 
 			// Check that each server received expected oneway messages
-			for i := 0; i < len(srvs); i++ {
+			for i := range srvs {
 				srvs[i].wg.Wait()
 				close(srvs[i].received)
 				expected := uint64(1)
 				for r := range srvs[i].received {
-					if expected != r.Num {
-						t.Errorf("%s(%d) = %d, expected %d", test.name, expected, r.Num, expected)
+					if expected != r.GetNum() {
+						t.Errorf("%s(%d) = %d, expected %d", test.name, expected, r.GetNum(), expected)
 					}
 					expected++
 				}
@@ -141,9 +139,9 @@ func TestMulticastPerNode(t *testing.T) {
 
 	// simple transformation function
 	f := func(msg *oneway.Request, id uint32) *oneway.Request {
-		return &oneway.Request{Num: add(msg.Num, id)}
+		return oneway.Request_builder{Num: add(msg.GetNum(), id)}.Build()
 	}
-	ignoreNodes := []int{}
+	var ignoreNodes []int
 	ignore := func(id uint32) bool {
 		for _, ignore := range ignoreNodes {
 			return id == uint32(ignore)
@@ -155,7 +153,7 @@ func TestMulticastPerNode(t *testing.T) {
 		if ignore(id) {
 			return nil
 		}
-		return &oneway.Request{Num: add(msg.Num, id)}
+		return oneway.Request_builder{Num: add(msg.GetNum(), id)}.Build()
 	}
 	tests := []struct {
 		name        string
@@ -188,7 +186,7 @@ func TestMulticastPerNode(t *testing.T) {
 			// set the ignoreNodes variable used by the ignore() function
 			ignoreNodes = test.ignoreNodes
 
-			for i := 0; i < len(srvs); i++ {
+			for i := range srvs {
 				if ignore(nodeIDs[i]) {
 					continue // don't check ignored nodes
 				}
@@ -196,7 +194,7 @@ func TestMulticastPerNode(t *testing.T) {
 			}
 
 			for c := 1; c <= test.calls; c++ {
-				in := &oneway.Request{Num: uint64(c)}
+				in := oneway.Request_builder{Num: uint64(c)}.Build()
 				if test.sendWait {
 					cfg.MulticastPerNode(context.Background(), in, test.f)
 				} else {
@@ -205,7 +203,7 @@ func TestMulticastPerNode(t *testing.T) {
 			}
 
 			// Check that each server received expected oneway messages
-			for i := 0; i < len(srvs); i++ {
+			for i := range srvs {
 				if ignore(nodeIDs[i]) {
 					continue // don't check ignored nodes
 				}
@@ -213,8 +211,8 @@ func TestMulticastPerNode(t *testing.T) {
 				srvs[i].wg.Wait()
 				close(srvs[i].received)
 				for r := range srvs[i].received {
-					if expected != r.Num {
-						t.Errorf("%s -> %d, expected %d, nodeID=%d, ignore=%t", test.name, r.Num, expected, nodeIDs[i], ignore(nodeIDs[i]))
+					if expected != r.GetNum() {
+						t.Errorf("%s -> %d, expected %d, nodeID=%d, ignore=%t", test.name, r.GetNum(), expected, nodeIDs[i], ignore(nodeIDs[i]))
 					}
 					expected++
 				}
@@ -230,16 +228,16 @@ func BenchmarkUnicast(b *testing.B) {
 		srv.benchmark = true
 	}
 	node := cfg.Nodes()[0]
-	in := &oneway.Request{Num: 0}
+	in := oneway.Request_builder{Num: 0}.Build()
 	b.Run("UnicastSendWaiting__", func(b *testing.B) {
 		for c := 1; c <= b.N; c++ {
-			in.Num = uint64(c)
+			in.SetNum(uint64(c))
 			node.Unicast(context.Background(), in)
 		}
 	})
 	b.Run("UnicastNoSendWaiting", func(b *testing.B) {
 		for c := 1; c <= b.N; c++ {
-			in.Num = uint64(c)
+			in.SetNum(uint64(c))
 			node.Unicast(context.Background(), in, gorums.WithNoSendWaiting())
 		}
 	})
@@ -251,16 +249,16 @@ func BenchmarkMulticast(b *testing.B) {
 	for _, srv := range srvs {
 		srv.benchmark = true
 	}
-	in := &oneway.Request{Num: 0}
+	in := oneway.Request_builder{Num: 0}.Build()
 	b.Run("MulticastSendWaiting__", func(b *testing.B) {
 		for c := 1; c <= b.N; c++ {
-			in.Num = uint64(c)
+			in.SetNum(uint64(c))
 			cfg.Multicast(context.Background(), in)
 		}
 	})
 	b.Run("MulticastNoSendWaiting", func(b *testing.B) {
 		for c := 1; c <= b.N; c++ {
-			in.Num = uint64(c)
+			in.SetNum(uint64(c))
 			cfg.Multicast(context.Background(), in, gorums.WithNoSendWaiting())
 		}
 	})
