@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
+// 	protoc-gen-gorums v0.9.0-devel+d97bfa83
 // 	protoc            v6.30.2
 // source: benchmark/benchmark.proto
 
@@ -8,18 +8,16 @@ package benchmark
 
 import (
 	context "context"
-	fmt "fmt"
 	gorums "github.com/relab/gorums"
 	encoding "google.golang.org/grpc/encoding"
-	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 func init() {
@@ -28,35 +26,15 @@ func init() {
 	}
 }
 
-// AsyncQuorumCall asynchronously invokes a quorum call on configuration c
-// and returns a AsyncEcho, which can be used to inspect the quorum call
-// reply and error when available.
-func (c *BenchmarkConfiguration) AsyncQuorumCall(ctx context.Context, in *Echo) *AsyncEcho {
-	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "benchmark.Benchmark.AsyncQuorumCall",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Echo, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Echo)
-		}
-		return c.qspec.AsyncQuorumCallQF(req.(*Echo), r)
-	}
-
-	fut := c.RawConfiguration.AsyncCall(ctx, cd)
-	return &AsyncEcho{fut}
-}
-
 // BenchmarkClient is the client interface for the Benchmark service.
 type BenchmarkClient interface {
-	StartServerBenchmark(ctx context.Context, in *StartRequest) (resp *StartResponse, err error)
-	StopServerBenchmark(ctx context.Context, in *StopRequest) (resp *Result, err error)
-	StartBenchmark(ctx context.Context, in *StartRequest) (resp *StartResponse, err error)
-	StopBenchmark(ctx context.Context, in *StopRequest) (resp *MemoryStatList, err error)
-	QuorumCall(ctx context.Context, in *Echo) (resp *Echo, err error)
-	AsyncQuorumCall(ctx context.Context, in *Echo) *AsyncEcho
-	SlowServer(ctx context.Context, in *Echo) (resp *Echo, err error)
+	StartServerBenchmark(ctx context.Context, in *StartRequest) gorums.Responses[*StartResponse]
+	StopServerBenchmark(ctx context.Context, in *StopRequest) gorums.Responses[*Result]
+	StartBenchmark(ctx context.Context, in *StartRequest) gorums.Responses[*StartResponse]
+	StopBenchmark(ctx context.Context, in *StopRequest) gorums.Responses[*MemoryStat]
+	QuorumCall(ctx context.Context, in *Echo) gorums.Responses[*Echo]
+	AsyncQuorumCall(ctx context.Context, in *Echo) gorums.Responses[*Echo]
+	SlowServer(ctx context.Context, in *Echo) gorums.Responses[*Echo]
 	Multicast(ctx context.Context, in *TimedMsg, opts ...gorums.CallOption)
 }
 
@@ -74,29 +52,47 @@ var _ BenchmarkNodeClient = (*BenchmarkNode)(nil)
 // procedure calls may be invoked.
 type BenchmarkConfiguration struct {
 	gorums.RawConfiguration
-	qspec BenchmarkQuorumSpec
-	nodes []*BenchmarkNode
 }
 
-// BenchmarkQuorumSpecFromRaw returns a new BenchmarkQuorumSpec from the given raw configuration and QuorumSpec.
+// NewBenchmarkConfiguration returns a configuration based on the provided list of nodes (required)
+// and an optional quorum specification. The QuorumSpec is necessary for call types that
+// must process replies. For configurations only used for unicast or multicast call types,
+// a QuorumSpec is not needed.
+// Nodes can be supplied using WithNodeMap or WithNodeList.
+// Using any other type of NodeListOption will not work.
+// The ManagerOption list controls how the nodes in the configuration are created.
+func NewBenchmarkConfiguration(cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *BenchmarkConfiguration, err error) {
+	c = &BenchmarkConfiguration{}
+	c.RawConfiguration, err = gorums.NewRawConfiguration(cfg, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return c, nil
+}
+
+// SubBenchmarkConfiguration allows for making a new Configuration from the
+// ManagerOption list and node list of another set of configurations,
+// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
+// A new configuration can also be created from an existing configuration,
+// using the And, WithNewNodes, Except, and WithoutNodes methods.
+func (c *BenchmarkConfiguration) SubBenchmarkConfiguration(cfg gorums.NodeListOption) (subCfg *BenchmarkConfiguration, err error) {
+	subCfg = &BenchmarkConfiguration{}
+	subCfg.RawConfiguration, err = c.SubRawConfiguration(cfg)
+	if err != nil {
+		return nil, err
+	}
+	return subCfg, nil
+}
+
+// BenchmarkConfigurationFromRaw returns a new BenchmarkConfiguration from the given raw configuration.
 //
 // This function may for example be used to "clone" a configuration but install a different QuorumSpec:
 //
 //	cfg1, err := mgr.NewConfiguration(qspec1, opts...)
 //	cfg2 := ConfigurationFromRaw(cfg1.RawConfig, qspec2)
-func BenchmarkConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec BenchmarkQuorumSpec) (*BenchmarkConfiguration, error) {
-	// return an error if qspec is nil.
-	if qspec == nil {
-		return nil, fmt.Errorf("config: missing required QuorumSpec")
-	}
+func BenchmarkConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*BenchmarkConfiguration, error) {
 	newCfg := &BenchmarkConfiguration{
 		RawConfiguration: rawCfg,
-		qspec:            qspec,
-	}
-	// initialize the nodes slice
-	newCfg.nodes = make([]*BenchmarkNode, newCfg.Size())
-	for i, n := range rawCfg {
-		newCfg.nodes[i] = &BenchmarkNode{n}
 	}
 	return newCfg, nil
 }
@@ -106,7 +102,24 @@ func BenchmarkConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec Benchma
 //
 // NOTE: mutating the returned slice is not supported.
 func (c *BenchmarkConfiguration) Nodes() []*BenchmarkNode {
-	return c.nodes
+	rawNodes := c.RawConfiguration.Nodes()
+	nodes := make([]*BenchmarkNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &BenchmarkNode{n}
+	}
+	return nodes
+}
+
+// AllNodes returns a slice of each available node of all subconfigurations. Sorted by node id.
+//
+// NOTE: mutating the returned slice is not supported.
+func (c *BenchmarkConfiguration) AllNodes() []*BenchmarkNode {
+	rawNodes := c.RawConfiguration.AllNodes()
+	nodes := make([]*BenchmarkNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &BenchmarkNode{n}
+	}
+	return nodes
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
@@ -120,61 +133,14 @@ func (c BenchmarkConfiguration) Except(rm *BenchmarkConfiguration) gorums.NodeLi
 	return c.RawConfiguration.Except(rm.RawConfiguration)
 }
 
-// BenchmarkManager maintains a connection pool of nodes on
-// which quorum calls can be performed.
-type BenchmarkManager struct {
-	*gorums.RawManager
-}
-
-// NewBenchmarkManager returns a new BenchmarkManager for managing connection to nodes added
-// to the manager. This function accepts manager options used to configure
-// various aspects of the manager.
-func NewBenchmarkManager(opts ...gorums.ManagerOption) *BenchmarkManager {
-	return &BenchmarkManager{
-		RawManager: gorums.NewRawManager(opts...),
-	}
-}
-
-// NewBenchmarkConfiguration returns a BenchmarkConfiguration based on the provided list of nodes (required)and a quorum specification.
-// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
-// A new configuration can also be created from an existing configuration,
-// using the And, WithNewNodes, Except, and WithoutNodes methods.
-func (m *BenchmarkManager) NewConfiguration(cfg gorums.NodeListOption, qspec BenchmarkQuorumSpec) (c *BenchmarkConfiguration, err error) {
-	c = &BenchmarkConfiguration{}
-	c.RawConfiguration, err = gorums.NewRawConfiguration(m.RawManager, cfg)
-	if err != nil {
-		return nil, err
-	}
-	// return an error if qspec is nil.
-	if qspec == nil {
-		return nil, fmt.Errorf("config: missing required BenchmarkQuorumSpec")
-	}
-	c.qspec = qspec
-	// initialize the nodes slice
-	c.nodes = make([]*BenchmarkNode, c.Size())
-	for i, n := range c.RawConfiguration {
-		c.nodes[i] = &BenchmarkNode{n}
-	}
-	return c, nil
-}
-
-// Nodes returns a slice of available nodes on this manager.
-// IDs are returned in the order they were added at creation of the manager.
-func (m *BenchmarkManager) Nodes() []*BenchmarkNode {
-	gorumsNodes := m.RawManager.Nodes()
-	nodes := make([]*BenchmarkNode, len(gorumsNodes))
-	for i, n := range gorumsNodes {
-		nodes[i] = &BenchmarkNode{n}
-	}
-	return nodes
-}
-
 // Reference imports to suppress errors if they are not otherwise used.
 var _ emptypb.Empty
 
-// Multicast is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
+// Multicast is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
 func (c *BenchmarkConfiguration) Multicast(ctx context.Context, in *TimedMsg, opts ...gorums.CallOption) {
+
 	cd := gorums.QuorumCallData{
 		Message: in,
 		Method:  "benchmark.Benchmark.Multicast",
@@ -188,189 +154,93 @@ type BenchmarkNode struct {
 	*gorums.RawNode
 }
 
-// BenchmarkQuorumSpec is the interface of quorum functions for Benchmark.
-type BenchmarkQuorumSpec interface {
-	gorums.ConfigOption
+// StartServerBenchmark is a quorum call invoked on each node in configuration c,
 
-	// StartServerBenchmarkQF is the quorum function for the StartServerBenchmark
-	// quorum call method. The in parameter is the request object
-	// supplied to the StartServerBenchmark method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *StartRequest'.
-	StartServerBenchmarkQF(in *StartRequest, replies map[uint32]*StartResponse) (*StartResponse, bool)
+// with the same argument in, and returns the responses as an iterator.
+func (c *BenchmarkConfiguration) StartServerBenchmark(ctx context.Context, in *StartRequest) gorums.Responses[*StartResponse] {
+	cd := gorums.QuorumCallData{
+		Message:      in,
+		Method:       "benchmark.Benchmark.StartServerBenchmark",
+		ServerStream: false,
+	}
 
-	// StopServerBenchmarkQF is the quorum function for the StopServerBenchmark
-	// quorum call method. The in parameter is the request object
-	// supplied to the StopServerBenchmark method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *StopRequest'.
-	StopServerBenchmarkQF(in *StopRequest, replies map[uint32]*Result) (*Result, bool)
-
-	// StartBenchmarkQF is the quorum function for the StartBenchmark
-	// quorum call method. The in parameter is the request object
-	// supplied to the StartBenchmark method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *StartRequest'.
-	StartBenchmarkQF(in *StartRequest, replies map[uint32]*StartResponse) (*StartResponse, bool)
-
-	// StopBenchmarkQF is the quorum function for the StopBenchmark
-	// quorum call method. The in parameter is the request object
-	// supplied to the StopBenchmark method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *StopRequest'.
-	StopBenchmarkQF(in *StopRequest, replies map[uint32]*MemoryStat) (*MemoryStatList, bool)
-
-	// QuorumCallQF is the quorum function for the QuorumCall
-	// quorum call method. The in parameter is the request object
-	// supplied to the QuorumCall method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *Echo'.
-	QuorumCallQF(in *Echo, replies map[uint32]*Echo) (*Echo, bool)
-
-	// AsyncQuorumCallQF is the quorum function for the AsyncQuorumCall
-	// asynchronous quorum call method. The in parameter is the request object
-	// supplied to the AsyncQuorumCall method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *Echo'.
-	AsyncQuorumCallQF(in *Echo, replies map[uint32]*Echo) (*Echo, bool)
-
-	// SlowServerQF is the quorum function for the SlowServer
-	// quorum call method. The in parameter is the request object
-	// supplied to the SlowServer method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *Echo'.
-	SlowServerQF(in *Echo, replies map[uint32]*Echo) (*Echo, bool)
+	return gorums.QuorumCall[*StartResponse](ctx, c.RawConfiguration, cd)
 }
 
-// StartServerBenchmark is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
-func (c *BenchmarkConfiguration) StartServerBenchmark(ctx context.Context, in *StartRequest) (resp *StartResponse, err error) {
+// StopServerBenchmark is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
+func (c *BenchmarkConfiguration) StopServerBenchmark(ctx context.Context, in *StopRequest) gorums.Responses[*Result] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "benchmark.Benchmark.StartServerBenchmark",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*StartResponse, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*StartResponse)
-		}
-		return c.qspec.StartServerBenchmarkQF(req.(*StartRequest), r)
+		Message:      in,
+		Method:       "benchmark.Benchmark.StopServerBenchmark",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*StartResponse), err
+	return gorums.QuorumCall[*Result](ctx, c.RawConfiguration, cd)
 }
 
-// StopServerBenchmark is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
-func (c *BenchmarkConfiguration) StopServerBenchmark(ctx context.Context, in *StopRequest) (resp *Result, err error) {
+// StartBenchmark is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
+func (c *BenchmarkConfiguration) StartBenchmark(ctx context.Context, in *StartRequest) gorums.Responses[*StartResponse] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "benchmark.Benchmark.StopServerBenchmark",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Result, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Result)
-		}
-		return c.qspec.StopServerBenchmarkQF(req.(*StopRequest), r)
+		Message:      in,
+		Method:       "benchmark.Benchmark.StartBenchmark",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*Result), err
+	return gorums.QuorumCall[*StartResponse](ctx, c.RawConfiguration, cd)
 }
 
-// StartBenchmark is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
-func (c *BenchmarkConfiguration) StartBenchmark(ctx context.Context, in *StartRequest) (resp *StartResponse, err error) {
+// StopBenchmark is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
+func (c *BenchmarkConfiguration) StopBenchmark(ctx context.Context, in *StopRequest) gorums.Responses[*MemoryStat] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "benchmark.Benchmark.StartBenchmark",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*StartResponse, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*StartResponse)
-		}
-		return c.qspec.StartBenchmarkQF(req.(*StartRequest), r)
+		Message:      in,
+		Method:       "benchmark.Benchmark.StopBenchmark",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*StartResponse), err
-}
-
-// StopBenchmark is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
-func (c *BenchmarkConfiguration) StopBenchmark(ctx context.Context, in *StopRequest) (resp *MemoryStatList, err error) {
-	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "benchmark.Benchmark.StopBenchmark",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*MemoryStat, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*MemoryStat)
-		}
-		return c.qspec.StopBenchmarkQF(req.(*StopRequest), r)
-	}
-
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*MemoryStatList), err
+	return gorums.QuorumCall[*MemoryStat](ctx, c.RawConfiguration, cd)
 }
 
 // benchmarks
-func (c *BenchmarkConfiguration) QuorumCall(ctx context.Context, in *Echo) (resp *Echo, err error) {
+func (c *BenchmarkConfiguration) QuorumCall(ctx context.Context, in *Echo) gorums.Responses[*Echo] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "benchmark.Benchmark.QuorumCall",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Echo, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Echo)
-		}
-		return c.qspec.QuorumCallQF(req.(*Echo), r)
+		Message:      in,
+		Method:       "benchmark.Benchmark.QuorumCall",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*Echo), err
+	return gorums.QuorumCall[*Echo](ctx, c.RawConfiguration, cd)
 }
 
-// SlowServer is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
-func (c *BenchmarkConfiguration) SlowServer(ctx context.Context, in *Echo) (resp *Echo, err error) {
+// AsyncQuorumCall is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
+func (c *BenchmarkConfiguration) AsyncQuorumCall(ctx context.Context, in *Echo) gorums.Responses[*Echo] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "benchmark.Benchmark.SlowServer",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Echo, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Echo)
-		}
-		return c.qspec.SlowServerQF(req.(*Echo), r)
+		Message:      in,
+		Method:       "benchmark.Benchmark.AsyncQuorumCall",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
+	return gorums.QuorumCall[*Echo](ctx, c.RawConfiguration, cd)
+}
+
+// SlowServer is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
+func (c *BenchmarkConfiguration) SlowServer(ctx context.Context, in *Echo) gorums.Responses[*Echo] {
+	cd := gorums.QuorumCallData{
+		Message:      in,
+		Method:       "benchmark.Benchmark.SlowServer",
+		ServerStream: false,
 	}
-	return res.(*Echo), err
+
+	return gorums.QuorumCall[*Echo](ctx, c.RawConfiguration, cd)
 }
 
 // Benchmark is the server-side API for the Benchmark Service
@@ -433,43 +303,4 @@ func RegisterBenchmarkServer(srv *gorums.Server, impl BenchmarkServer) {
 		defer ctx.Release()
 		impl.Multicast(ctx, req)
 	})
-}
-
-type internalEcho struct {
-	nid   uint32
-	reply *Echo
-	err   error
-}
-
-type internalMemoryStat struct {
-	nid   uint32
-	reply *MemoryStat
-	err   error
-}
-
-type internalResult struct {
-	nid   uint32
-	reply *Result
-	err   error
-}
-
-type internalStartResponse struct {
-	nid   uint32
-	reply *StartResponse
-	err   error
-}
-
-// AsyncEcho is a async object for processing replies.
-type AsyncEcho struct {
-	*gorums.Async
-}
-
-// Get returns the reply and any error associated with the called method.
-// The method blocks until a reply or error is available.
-func (f *AsyncEcho) Get() (*Echo, error) {
-	resp, err := f.Async.Get()
-	if err != nil {
-		return nil, err
-	}
-	return resp.(*Echo), err
 }
