@@ -1,50 +1,67 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
+// 	protoc-gen-gorums v0.9.0-devel+d97bfa83
 // 	protoc            v6.30.2
 // source: zorums.proto
 
 package dev
 
 import (
-	fmt "fmt"
 	gorums "github.com/relab/gorums"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // A ZorumsServiceConfiguration represents a static set of nodes on which quorum remote
 // procedure calls may be invoked.
 type ZorumsServiceConfiguration struct {
 	gorums.RawConfiguration
-	qspec ZorumsServiceQuorumSpec
-	nodes []*ZorumsServiceNode
 }
 
-// ZorumsServiceQuorumSpecFromRaw returns a new ZorumsServiceQuorumSpec from the given raw configuration and QuorumSpec.
+// NewZorumsServiceConfiguration returns a configuration based on the provided list of nodes (required)
+// and an optional quorum specification. The QuorumSpec is necessary for call types that
+// must process replies. For configurations only used for unicast or multicast call types,
+// a QuorumSpec is not needed.
+// Nodes can be supplied using WithNodeMap or WithNodeList.
+// Using any other type of NodeListOption will not work.
+// The ManagerOption list controls how the nodes in the configuration are created.
+func NewZorumsServiceConfiguration(cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *ZorumsServiceConfiguration, err error) {
+	c = &ZorumsServiceConfiguration{}
+	c.RawConfiguration, err = gorums.NewRawConfiguration(cfg, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return c, nil
+}
+
+// SubZorumsServiceConfiguration allows for making a new Configuration from the
+// ManagerOption list and node list of another set of configurations,
+// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
+// A new configuration can also be created from an existing configuration,
+// using the And, WithNewNodes, Except, and WithoutNodes methods.
+func (c *ZorumsServiceConfiguration) SubZorumsServiceConfiguration(cfg gorums.NodeListOption) (subCfg *ZorumsServiceConfiguration, err error) {
+	subCfg = &ZorumsServiceConfiguration{}
+	subCfg.RawConfiguration, err = c.SubRawConfiguration(cfg)
+	if err != nil {
+		return nil, err
+	}
+	return subCfg, nil
+}
+
+// ZorumsServiceConfigurationFromRaw returns a new ZorumsServiceConfiguration from the given raw configuration.
 //
 // This function may for example be used to "clone" a configuration but install a different QuorumSpec:
 //
 //	cfg1, err := mgr.NewConfiguration(qspec1, opts...)
 //	cfg2 := ConfigurationFromRaw(cfg1.RawConfig, qspec2)
-func ZorumsServiceConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec ZorumsServiceQuorumSpec) (*ZorumsServiceConfiguration, error) {
-	// return an error if qspec is nil.
-	if qspec == nil {
-		return nil, fmt.Errorf("config: missing required QuorumSpec")
-	}
+func ZorumsServiceConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*ZorumsServiceConfiguration, error) {
 	newCfg := &ZorumsServiceConfiguration{
 		RawConfiguration: rawCfg,
-		qspec:            qspec,
-	}
-	// initialize the nodes slice
-	newCfg.nodes = make([]*ZorumsServiceNode, newCfg.Size())
-	for i, n := range rawCfg {
-		newCfg.nodes[i] = &ZorumsServiceNode{n}
 	}
 	return newCfg, nil
 }
@@ -54,7 +71,24 @@ func ZorumsServiceConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec Zor
 //
 // NOTE: mutating the returned slice is not supported.
 func (c *ZorumsServiceConfiguration) Nodes() []*ZorumsServiceNode {
-	return c.nodes
+	rawNodes := c.RawConfiguration.Nodes()
+	nodes := make([]*ZorumsServiceNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &ZorumsServiceNode{n}
+	}
+	return nodes
+}
+
+// AllNodes returns a slice of each available node of all subconfigurations. Sorted by node id.
+//
+// NOTE: mutating the returned slice is not supported.
+func (c *ZorumsServiceConfiguration) AllNodes() []*ZorumsServiceNode {
+	rawNodes := c.RawConfiguration.AllNodes()
+	nodes := make([]*ZorumsServiceNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &ZorumsServiceNode{n}
+	}
+	return nodes
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
@@ -68,27 +102,51 @@ func (c ZorumsServiceConfiguration) Except(rm *ZorumsServiceConfiguration) gorum
 	return c.RawConfiguration.Except(rm.RawConfiguration)
 }
 
-// A ZorumsNoQspecServiceConfiguration represents a static set of nodes on which quorum remote
+// A ZorumsNoQCServiceConfiguration represents a static set of nodes on which quorum remote
 // procedure calls may be invoked.
-type ZorumsNoQspecServiceConfiguration struct {
+type ZorumsNoQCServiceConfiguration struct {
 	gorums.RawConfiguration
-	nodes []*ZorumsNoQspecServiceNode
 }
 
-// ZorumsNoQspecServiceQuorumSpecFromRaw returns a new ZorumsNoQspecServiceQuorumSpec from the given raw configuration.
+// NewZorumsNoQCServiceConfiguration returns a configuration based on the provided list of nodes (required)
+// and an optional quorum specification. The QuorumSpec is necessary for call types that
+// must process replies. For configurations only used for unicast or multicast call types,
+// a QuorumSpec is not needed.
+// Nodes can be supplied using WithNodeMap or WithNodeList.
+// Using any other type of NodeListOption will not work.
+// The ManagerOption list controls how the nodes in the configuration are created.
+func NewZorumsNoQCServiceConfiguration(cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *ZorumsNoQCServiceConfiguration, err error) {
+	c = &ZorumsNoQCServiceConfiguration{}
+	c.RawConfiguration, err = gorums.NewRawConfiguration(cfg, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return c, nil
+}
+
+// SubZorumsNoQCServiceConfiguration allows for making a new Configuration from the
+// ManagerOption list and node list of another set of configurations,
+// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
+// A new configuration can also be created from an existing configuration,
+// using the And, WithNewNodes, Except, and WithoutNodes methods.
+func (c *ZorumsNoQCServiceConfiguration) SubZorumsNoQCServiceConfiguration(cfg gorums.NodeListOption) (subCfg *ZorumsNoQCServiceConfiguration, err error) {
+	subCfg = &ZorumsNoQCServiceConfiguration{}
+	subCfg.RawConfiguration, err = c.SubRawConfiguration(cfg)
+	if err != nil {
+		return nil, err
+	}
+	return subCfg, nil
+}
+
+// ZorumsNoQCServiceConfigurationFromRaw returns a new ZorumsNoQCServiceConfiguration from the given raw configuration.
 //
 // This function may for example be used to "clone" a configuration but install a different QuorumSpec:
 //
 //	cfg1, err := mgr.NewConfiguration(qspec1, opts...)
 //	cfg2 := ConfigurationFromRaw(cfg1.RawConfig, qspec2)
-func ZorumsNoQspecServiceConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*ZorumsNoQspecServiceConfiguration, error) {
-	newCfg := &ZorumsNoQspecServiceConfiguration{
+func ZorumsNoQCServiceConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*ZorumsNoQCServiceConfiguration, error) {
+	newCfg := &ZorumsNoQCServiceConfiguration{
 		RawConfiguration: rawCfg,
-	}
-	// initialize the nodes slice
-	newCfg.nodes = make([]*ZorumsNoQspecServiceNode, newCfg.Size())
-	for i, n := range rawCfg {
-		newCfg.nodes[i] = &ZorumsNoQspecServiceNode{n}
 	}
 	return newCfg, nil
 }
@@ -97,17 +155,34 @@ func ZorumsNoQspecServiceConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*
 // order as they were provided in the creation of the Manager.
 //
 // NOTE: mutating the returned slice is not supported.
-func (c *ZorumsNoQspecServiceConfiguration) Nodes() []*ZorumsNoQspecServiceNode {
-	return c.nodes
+func (c *ZorumsNoQCServiceConfiguration) Nodes() []*ZorumsNoQCServiceNode {
+	rawNodes := c.RawConfiguration.Nodes()
+	nodes := make([]*ZorumsNoQCServiceNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &ZorumsNoQCServiceNode{n}
+	}
+	return nodes
+}
+
+// AllNodes returns a slice of each available node of all subconfigurations. Sorted by node id.
+//
+// NOTE: mutating the returned slice is not supported.
+func (c *ZorumsNoQCServiceConfiguration) AllNodes() []*ZorumsNoQCServiceNode {
+	rawNodes := c.RawConfiguration.AllNodes()
+	nodes := make([]*ZorumsNoQCServiceNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &ZorumsNoQCServiceNode{n}
+	}
+	return nodes
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
-func (c ZorumsNoQspecServiceConfiguration) And(d *ZorumsNoQspecServiceConfiguration) gorums.NodeListOption {
+func (c ZorumsNoQCServiceConfiguration) And(d *ZorumsNoQCServiceConfiguration) gorums.NodeListOption {
 	return c.RawConfiguration.And(d.RawConfiguration)
 }
 
 // Except returns a NodeListOption that can be used to create a new configuration
 // from c without the nodes in rm.
-func (c ZorumsNoQspecServiceConfiguration) Except(rm *ZorumsNoQspecServiceConfiguration) gorums.NodeListOption {
+func (c ZorumsNoQCServiceConfiguration) Except(rm *ZorumsNoQCServiceConfiguration) gorums.NodeListOption {
 	return c.RawConfiguration.Except(rm.RawConfiguration)
 }
