@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
+// 	protoc-gen-gorums v0.9.0-devel+d97bfa83
 // 	protoc            v6.30.2
 // source: zorums.proto
 
@@ -9,19 +9,20 @@ package dev
 import (
 	context "context"
 	gorums "github.com/relab/gorums"
-	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
+	proto "google.golang.org/protobuf/proto"
 	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // Multicast plain. Response type is not needed here.
 func (c *ZorumsServiceConfiguration) Multicast(ctx context.Context, in *Request, opts ...gorums.CallOption) {
+
 	cd := gorums.QuorumCallData{
 		Message: in,
 		Method:  "dev.ZorumsService.Multicast",
@@ -32,12 +33,13 @@ func (c *ZorumsServiceConfiguration) Multicast(ctx context.Context, in *Request,
 
 // MulticastPerNodeArg with per_node_arg option.
 func (c *ZorumsServiceConfiguration) MulticastPerNodeArg(ctx context.Context, in *Request, f func(*Request, uint32) *Request, opts ...gorums.CallOption) {
+
 	cd := gorums.QuorumCallData{
 		Message: in,
 		Method:  "dev.ZorumsService.MulticastPerNodeArg",
 	}
 
-	cd.PerNodeArgFn = func(req protoreflect.ProtoMessage, nid uint32) protoreflect.ProtoMessage {
+	cd.PerNodeArgFn = func(req proto.Message, nid uint32) proto.Message {
 		return f(req.(*Request), nid)
 	}
 
@@ -46,6 +48,7 @@ func (c *ZorumsServiceConfiguration) MulticastPerNodeArg(ctx context.Context, in
 
 // Multicast2 is testing whether multiple streams work.
 func (c *ZorumsServiceConfiguration) Multicast2(ctx context.Context, in *Request, opts ...gorums.CallOption) {
+
 	cd := gorums.QuorumCallData{
 		Message: in,
 		Method:  "dev.ZorumsService.Multicast2",
@@ -59,6 +62,7 @@ var _ emptypb.Empty
 
 // Multicast3 is testing imported message type.
 func (c *ZorumsServiceConfiguration) Multicast3(ctx context.Context, in *Request, opts ...gorums.CallOption) {
+
 	cd := gorums.QuorumCallData{
 		Message: in,
 		Method:  "dev.ZorumsService.Multicast3",
@@ -72,6 +76,7 @@ var _ emptypb.Empty
 
 // Multicast4 is testing imported message type.
 func (c *ZorumsServiceConfiguration) Multicast4(ctx context.Context, in *emptypb.Empty, opts ...gorums.CallOption) {
+
 	cd := gorums.QuorumCallData{
 		Message: in,
 		Method:  "dev.ZorumsService.Multicast4",
@@ -81,23 +86,25 @@ func (c *ZorumsServiceConfiguration) Multicast4(ctx context.Context, in *emptypb
 }
 
 // Multicast plain. Response type is not needed here.
-func (c *ZorumsNoQspecServiceConfiguration) Multicast(ctx context.Context, in *Request, opts ...gorums.CallOption) {
+func (c *ZorumsNoQCServiceConfiguration) Multicast(ctx context.Context, in *Request, opts ...gorums.CallOption) {
+
 	cd := gorums.QuorumCallData{
 		Message: in,
-		Method:  "dev.ZorumsNoQspecService.Multicast",
+		Method:  "dev.ZorumsNoQCService.Multicast",
 	}
 
 	c.RawConfiguration.Multicast(ctx, cd, opts...)
 }
 
 // MulticastPerNodeArg with per_node_arg option.
-func (c *ZorumsNoQspecServiceConfiguration) MulticastPerNodeArg(ctx context.Context, in *Request, f func(*Request, uint32) *Request, opts ...gorums.CallOption) {
+func (c *ZorumsNoQCServiceConfiguration) MulticastPerNodeArg(ctx context.Context, in *Request, f func(*Request, uint32) *Request, opts ...gorums.CallOption) {
+
 	cd := gorums.QuorumCallData{
 		Message: in,
-		Method:  "dev.ZorumsNoQspecService.MulticastPerNodeArg",
+		Method:  "dev.ZorumsNoQCService.MulticastPerNodeArg",
 	}
 
-	cd.PerNodeArgFn = func(req protoreflect.ProtoMessage, nid uint32) protoreflect.ProtoMessage {
+	cd.PerNodeArgFn = func(req proto.Message, nid uint32) proto.Message {
 		return f(req.(*Request), nid)
 	}
 
@@ -105,10 +112,11 @@ func (c *ZorumsNoQspecServiceConfiguration) MulticastPerNodeArg(ctx context.Cont
 }
 
 // Multicast2 is testing whether multiple streams work.
-func (c *ZorumsNoQspecServiceConfiguration) Multicast2(ctx context.Context, in *Request, opts ...gorums.CallOption) {
+func (c *ZorumsNoQCServiceConfiguration) Multicast2(ctx context.Context, in *Request, opts ...gorums.CallOption) {
+
 	cd := gorums.QuorumCallData{
 		Message: in,
-		Method:  "dev.ZorumsNoQspecService.Multicast2",
+		Method:  "dev.ZorumsNoQCService.Multicast2",
 	}
 
 	c.RawConfiguration.Multicast(ctx, cd, opts...)
@@ -118,10 +126,11 @@ func (c *ZorumsNoQspecServiceConfiguration) Multicast2(ctx context.Context, in *
 var _ emptypb.Empty
 
 // Multicast3 is testing imported message type.
-func (c *ZorumsNoQspecServiceConfiguration) Multicast3(ctx context.Context, in *Request, opts ...gorums.CallOption) {
+func (c *ZorumsNoQCServiceConfiguration) Multicast3(ctx context.Context, in *Request, opts ...gorums.CallOption) {
+
 	cd := gorums.QuorumCallData{
 		Message: in,
-		Method:  "dev.ZorumsNoQspecService.Multicast3",
+		Method:  "dev.ZorumsNoQCService.Multicast3",
 	}
 
 	c.RawConfiguration.Multicast(ctx, cd, opts...)
@@ -131,10 +140,11 @@ func (c *ZorumsNoQspecServiceConfiguration) Multicast3(ctx context.Context, in *
 var _ emptypb.Empty
 
 // Multicast4 is testing imported message type.
-func (c *ZorumsNoQspecServiceConfiguration) Multicast4(ctx context.Context, in *emptypb.Empty, opts ...gorums.CallOption) {
+func (c *ZorumsNoQCServiceConfiguration) Multicast4(ctx context.Context, in *emptypb.Empty, opts ...gorums.CallOption) {
+
 	cd := gorums.QuorumCallData{
 		Message: in,
-		Method:  "dev.ZorumsNoQspecService.Multicast4",
+		Method:  "dev.ZorumsNoQCService.Multicast4",
 	}
 
 	c.RawConfiguration.Multicast(ctx, cd, opts...)
