@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
+// 	protoc-gen-gorums v0.9.0-devel+d97bfa83
 // 	protoc            v6.30.2
 // source: zorums.proto
 
@@ -12,9 +12,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // ZorumsServiceNode holds the node specific methods for the ZorumsService service.
@@ -22,7 +22,7 @@ type ZorumsServiceNode struct {
 	*gorums.RawNode
 }
 
-// ZorumsNoQspecServiceNode holds the node specific methods for the ZorumsNoQspecService service.
-type ZorumsNoQspecServiceNode struct {
+// ZorumsNoQCServiceNode holds the node specific methods for the ZorumsNoQCService service.
+type ZorumsNoQCServiceNode struct {
 	*gorums.RawNode
 }
