@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
+// 	protoc-gen-gorums v0.9.0-devel+d97bfa83
 // 	protoc            v6.30.2
 // source: zorums.proto
 
@@ -13,9 +13,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // GRPCCall plain gRPC call; testing that Gorums can ignore these, but that
@@ -35,10 +35,10 @@ func (n *ZorumsServiceNode) GRPCCall(ctx context.Context, in *Request) (resp *Re
 
 // GRPCCall plain gRPC call; testing that Gorums can ignore these, but that
 // they are added to the _grpc.pb.go generated file.
-func (n *ZorumsNoQspecServiceNode) GRPCCall(ctx context.Context, in *Request) (resp *Response, err error) {
+func (n *ZorumsNoQCServiceNode) GRPCCall(ctx context.Context, in *Request) (resp *Response, err error) {
 	cd := gorums.CallData{
 		Message: in,
-		Method:  "dev.ZorumsNoQspecService.GRPCCall",
+		Method:  "dev.ZorumsNoQCService.GRPCCall",
 	}
 
 	res, err := n.RawNode.RPCCall(ctx, cd)
