@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
+// 	protoc-gen-gorums v0.9.0-devel+d97bfa83
 // 	protoc            v6.30.2
 // source: zorums.proto
 
@@ -14,9 +14,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // Unicast is a one-way call; no replies are processed.
@@ -43,10 +43,10 @@ func (n *ZorumsServiceNode) Unicast2(ctx context.Context, in *Request, opts ...g
 }
 
 // Unicast is a one-way call; no replies are processed.
-func (n *ZorumsNoQspecServiceNode) Unicast(ctx context.Context, in *Request, opts ...gorums.CallOption) {
+func (n *ZorumsNoQCServiceNode) Unicast(ctx context.Context, in *Request, opts ...gorums.CallOption) {
 	cd := gorums.CallData{
 		Message: in,
-		Method:  "dev.ZorumsNoQspecService.Unicast",
+		Method:  "dev.ZorumsNoQCService.Unicast",
 	}
 
 	n.RawNode.Unicast(ctx, cd, opts...)
@@ -56,10 +56,10 @@ func (n *ZorumsNoQspecServiceNode) Unicast(ctx context.Context, in *Request, opt
 var _ emptypb.Empty
 
 // Unicast2 is a one-way call; no replies are processed.
-func (n *ZorumsNoQspecServiceNode) Unicast2(ctx context.Context, in *Request, opts ...gorums.CallOption) {
+func (n *ZorumsNoQCServiceNode) Unicast2(ctx context.Context, in *Request, opts ...gorums.CallOption) {
 	cd := gorums.CallData{
 		Message: in,
-		Method:  "dev.ZorumsNoQspecService.Unicast2",
+		Method:  "dev.ZorumsNoQCService.Unicast2",
 	}
 
 	n.RawNode.Unicast(ctx, cd, opts...)
