@@ -1,36 +1,36 @@
 package gengorums
 
 var correctableCallComment = `
-{{$comments := .Method.Comments.Leading}}
-{{if ne $comments ""}}
-{{$comments -}}
-{{else}}
-{{if hasPerNodeArg .Method}}
-// {{$method}} asynchronously invokes a correctable quorum call on each node
-// in configuration c, with the argument returned by the provided function f
-// and returns a {{$correctableOut}}, which can be used to inspect
-// the reply and error when available.
-// The provide per node function f takes the provided {{$in}} argument
-// and returns an {{$out}} object to be passed to the given nodeID.
-// The per node function f should be thread-safe.
-{{else}}
-// {{$method}} asynchronously invokes a correctable quorum call on each node
-// in configuration c and returns a {{$correctableOut}}, which can be used
-// to inspect any replies or errors when available.
-{{if correctableStream .Method -}}
-// This method supports server-side preliminary replies (correctable stream).
-{{end -}}
-{{end -}}
-{{end -}}
+	{{$comments := .Method.Comments.Leading}}
+	{{if ne $comments ""}}
+	{{$comments -}}
+	{{else}}
+	{{if hasPerNodeArg .Method}}
+	// {{$method}} asynchronously invokes a correctable quorum call on each node
+	// in configuration c, with the argument returned by the provided function f
+	// and returns a {{$correctableOut}}, which can be used to inspect
+	// the reply and error when available.
+	// The provide per node function f takes the provided {{$in}} argument
+	// and returns an {{$out}} object to be passed to the given nodeID.
+	// The per node function f should be thread-safe.
+	{{else}}
+	// {{$method}} asynchronously invokes a correctable quorum call on each node
+	// in configuration c and returns a {{$correctableOut}}, which can be used
+	// to inspect any replies or errors when available.
+	{{if correctableStream .Method -}}
+	// This method supports server-side preliminary replies (correctable stream).
+	{{end -}}
+	{{end -}}
+	{{end -}}
 `
 
 var correctableVar = `
-{{$correctableOut := outType .Method $customOut}}
-{{$protoMessage := use "protoreflect.ProtoMessage" .GenFile}}
-{{$callData := use "gorums.CorrectableCallData" .GenFile}}
-{{$genFile := .GenFile}}
-{{$unexportMethod := unexport .Method.GoName}}
-{{$context := use "context.Context" .GenFile}}
+	{{$correctableOut := outType .Method $customOut}}
+	{{$protoMessage := use "protoreflect.ProtoMessage" .GenFile}}
+	{{$callData := use "gorums.CorrectableCallData" .GenFile}}
+	{{$genFile := .GenFile}}
+	{{$unexportMethod := unexport .Method.GoName}}
+	{{$context := use "context.Context" .GenFile}}
 `
 
 var correctableSignature = `func (c *{{$configurationName}}) {{$method}}(` +
@@ -50,11 +50,11 @@ var correctableBody = `	cd := {{$callData}}{
 		}
 		return c.qspec.{{$method}}QF(req.(*{{$in}}), r)
 	}
-{{- if hasPerNodeArg .Method}}
-	cd.PerNodeArgFn = func(req {{$protoMessage}}, nid uint32) {{$protoMessage}} {
-		return f(req.(*{{$in}}), nid)
-	}
-{{- end}}
+	{{- if hasPerNodeArg .Method}}
+		cd.PerNodeArgFn = func(req {{$protoMessage}}, nid uint32) {{$protoMessage}} {
+			return f(req.(*{{$in}}), nid)
+		}
+	{{- end}}
 
 	corr := c.RawConfiguration.CorrectableCall(ctx, cd)
 	return &{{$correctableOut}}{corr}
