@@ -7,12 +7,15 @@ var commonVariables = `
 {{$in := in .GenFile .Method}}
 {{$out := out .GenFile .Method}}
 {{$intOut := internalOut $out}}
-{{$customOut := customOut .GenFile .Method}}
-{{$customOutField := field $customOut}}
 {{$unexportOutput := unexport .Method.Output.GoIdent.GoName}}
 {{$serviceName := .Method.Parent.GoName}}
 {{$nodeName := printf "%sNode" $serviceName}}
 {{$configurationName := printf "%sConfiguration" $serviceName}}
+`
+
+// Common variables used in several template functions.
+var quorumCallVariables = `
+{{- $iterator := use "gorums.Responses" .GenFile}}
 `
 
 var quorumCallComment = `
@@ -20,15 +23,18 @@ var quorumCallComment = `
 {{if ne $comments ""}}
 {{$comments -}}
 {{else}}
-{{if hasPerNodeArg .Method}}
 // {{$method}} is a quorum call invoked on each node in configuration c,
-// with the argument returned by the provided function f, and returns the combined result.
+{{if hasPerNodeArg .Method}}
+// with the argument returned by the provided function f,
+// it returns the responses as an iterator.
 // The per node function f receives a copy of the {{$in}} request argument and
 // returns a {{$in}} manipulated to be passed to the given nodeID.
 // The function f must be thread-safe.
 {{else}}
-// {{$method}} is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
+// with the same argument in, and returns the responses as an iterator.
+{{end -}}
+{{if isStream .Method}}
+// This is a streaming quorum call, so each can respond with any amount of responses.
 {{end -}}
 {{end -}}
 `
@@ -36,11 +42,10 @@ var quorumCallComment = `
 var quorumCallSignature = `func (c *{{$configurationName}}) {{$method}}(` +
 	`ctx {{$context}}, in *{{$in}}` +
 	`{{perNodeFnType .GenFile .Method ", f"}})` +
-	`(resp *{{$customOut}}, err error) {
+	`{{$iterator}}[*{{$out}}] {
 `
 
 var qcVar = `
-{{$protoMessage := use "protoreflect.ProtoMessage" .GenFile}}
 {{$callData := use "gorums.QuorumCallData" .GenFile}}
 {{$genFile := .GenFile}}
 {{$unexportMethod := unexport .Method.GoName}}
@@ -50,29 +55,21 @@ var qcVar = `
 var quorumCallBody = `	cd := {{$callData}}{
 		Message: in,
 		Method:  "{{$fullName}}",
-	}
-	cd.QuorumFunction = func(req {{$protoMessage}}, replies map[uint32]{{$protoMessage}}) ({{$protoMessage}}, bool) {
-		r := make(map[uint32]*{{$out}}, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*{{$out}})
-		}
-		return c.qspec.{{$method}}QF(req.(*{{$in}}), r)
+		ServerStream: {{isStream .Method}},
 	}
 {{- if hasPerNodeArg .Method}}
+	{{- $protoMessage := use "proto.Message" .GenFile}}
 	cd.PerNodeArgFn = func(req {{$protoMessage}}, nid uint32) {{$protoMessage}} {
 		return f(req.(*{{$in}}), nid)
 	}
 {{- end}}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*{{$customOut}}), err
+	return gorums.QuorumCall[*{{$out}}](ctx, c.RawConfiguration, cd)
 }
 `
 
 var quorumCall = commonVariables +
+	quorumCallVariables +
 	qcVar +
 	quorumCallComment +
 	quorumCallSignature +
