@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
+// 	protoc-gen-gorums v0.9.0-devel+d97bfa83
 // 	protoc            v6.30.2
 // source: storage/proto/storage.proto
 
@@ -8,18 +8,16 @@ package proto
 
 import (
 	context "context"
-	fmt "fmt"
 	gorums "github.com/relab/gorums"
 	encoding "google.golang.org/grpc/encoding"
-	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 func init() {
@@ -30,8 +28,8 @@ func init() {
 
 // StorageClient is the client interface for the Storage service.
 type StorageClient interface {
-	ReadQC(ctx context.Context, in *ReadRequest) (resp *ReadResponse, err error)
-	WriteQC(ctx context.Context, in *WriteRequest) (resp *WriteResponse, err error)
+	ReadQC(ctx context.Context, in *ReadRequest) gorums.Responses[*ReadResponse]
+	WriteQC(ctx context.Context, in *WriteRequest) gorums.Responses[*WriteResponse]
 	WriteMulticast(ctx context.Context, in *WriteRequest, opts ...gorums.CallOption)
 }
 
@@ -51,29 +49,47 @@ var _ StorageNodeClient = (*StorageNode)(nil)
 // procedure calls may be invoked.
 type StorageConfiguration struct {
 	gorums.RawConfiguration
-	qspec StorageQuorumSpec
-	nodes []*StorageNode
 }
 
-// StorageQuorumSpecFromRaw returns a new StorageQuorumSpec from the given raw configuration and QuorumSpec.
+// NewStorageConfiguration returns a configuration based on the provided list of nodes (required)
+// and an optional quorum specification. The QuorumSpec is necessary for call types that
+// must process replies. For configurations only used for unicast or multicast call types,
+// a QuorumSpec is not needed.
+// Nodes can be supplied using WithNodeMap or WithNodeList.
+// Using any other type of NodeListOption will not work.
+// The ManagerOption list controls how the nodes in the configuration are created.
+func NewStorageConfiguration(cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *StorageConfiguration, err error) {
+	c = &StorageConfiguration{}
+	c.RawConfiguration, err = gorums.NewRawConfiguration(cfg, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return c, nil
+}
+
+// SubStorageConfiguration allows for making a new Configuration from the
+// ManagerOption list and node list of another set of configurations,
+// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
+// A new configuration can also be created from an existing configuration,
+// using the And, WithNewNodes, Except, and WithoutNodes methods.
+func (c *StorageConfiguration) SubStorageConfiguration(cfg gorums.NodeListOption) (subCfg *StorageConfiguration, err error) {
+	subCfg = &StorageConfiguration{}
+	subCfg.RawConfiguration, err = c.SubRawConfiguration(cfg)
+	if err != nil {
+		return nil, err
+	}
+	return subCfg, nil
+}
+
+// StorageConfigurationFromRaw returns a new StorageConfiguration from the given raw configuration.
 //
 // This function may for example be used to "clone" a configuration but install a different QuorumSpec:
 //
 //	cfg1, err := mgr.NewConfiguration(qspec1, opts...)
 //	cfg2 := ConfigurationFromRaw(cfg1.RawConfig, qspec2)
-func StorageConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec StorageQuorumSpec) (*StorageConfiguration, error) {
-	// return an error if qspec is nil.
-	if qspec == nil {
-		return nil, fmt.Errorf("config: missing required QuorumSpec")
-	}
+func StorageConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*StorageConfiguration, error) {
 	newCfg := &StorageConfiguration{
 		RawConfiguration: rawCfg,
-		qspec:            qspec,
-	}
-	// initialize the nodes slice
-	newCfg.nodes = make([]*StorageNode, newCfg.Size())
-	for i, n := range rawCfg {
-		newCfg.nodes[i] = &StorageNode{n}
 	}
 	return newCfg, nil
 }
@@ -83,7 +99,24 @@ func StorageConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec StorageQu
 //
 // NOTE: mutating the returned slice is not supported.
 func (c *StorageConfiguration) Nodes() []*StorageNode {
-	return c.nodes
+	rawNodes := c.RawConfiguration.Nodes()
+	nodes := make([]*StorageNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &StorageNode{n}
+	}
+	return nodes
+}
+
+// AllNodes returns a slice of each available node of all subconfigurations. Sorted by node id.
+//
+// NOTE: mutating the returned slice is not supported.
+func (c *StorageConfiguration) AllNodes() []*StorageNode {
+	rawNodes := c.RawConfiguration.AllNodes()
+	nodes := make([]*StorageNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &StorageNode{n}
+	}
+	return nodes
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
@@ -97,61 +130,14 @@ func (c StorageConfiguration) Except(rm *StorageConfiguration) gorums.NodeListOp
 	return c.RawConfiguration.Except(rm.RawConfiguration)
 }
 
-// StorageManager maintains a connection pool of nodes on
-// which quorum calls can be performed.
-type StorageManager struct {
-	*gorums.RawManager
-}
-
-// NewStorageManager returns a new StorageManager for managing connection to nodes added
-// to the manager. This function accepts manager options used to configure
-// various aspects of the manager.
-func NewStorageManager(opts ...gorums.ManagerOption) *StorageManager {
-	return &StorageManager{
-		RawManager: gorums.NewRawManager(opts...),
-	}
-}
-
-// NewStorageConfiguration returns a StorageConfiguration based on the provided list of nodes (required)and a quorum specification.
-// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
-// A new configuration can also be created from an existing configuration,
-// using the And, WithNewNodes, Except, and WithoutNodes methods.
-func (m *StorageManager) NewConfiguration(cfg gorums.NodeListOption, qspec StorageQuorumSpec) (c *StorageConfiguration, err error) {
-	c = &StorageConfiguration{}
-	c.RawConfiguration, err = gorums.NewRawConfiguration(m.RawManager, cfg)
-	if err != nil {
-		return nil, err
-	}
-	// return an error if qspec is nil.
-	if qspec == nil {
-		return nil, fmt.Errorf("config: missing required StorageQuorumSpec")
-	}
-	c.qspec = qspec
-	// initialize the nodes slice
-	c.nodes = make([]*StorageNode, c.Size())
-	for i, n := range c.RawConfiguration {
-		c.nodes[i] = &StorageNode{n}
-	}
-	return c, nil
-}
-
-// Nodes returns a slice of available nodes on this manager.
-// IDs are returned in the order they were added at creation of the manager.
-func (m *StorageManager) Nodes() []*StorageNode {
-	gorumsNodes := m.RawManager.Nodes()
-	nodes := make([]*StorageNode, len(gorumsNodes))
-	for i, n := range gorumsNodes {
-		nodes[i] = &StorageNode{n}
-	}
-	return nodes
-}
-
 // Reference imports to suppress errors if they are not otherwise used.
 var _ emptypb.Empty
 
-// WriteMulticast is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
+// WriteMulticast is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
 func (c *StorageConfiguration) WriteMulticast(ctx context.Context, in *WriteRequest, opts ...gorums.CallOption) {
+
 	cd := gorums.QuorumCallData{
 		Message: in,
 		Method:  "storage.Storage.WriteMulticast",
@@ -165,67 +151,28 @@ type StorageNode struct {
 	*gorums.RawNode
 }
 
-// StorageQuorumSpec is the interface of quorum functions for Storage.
-type StorageQuorumSpec interface {
-	gorums.ConfigOption
-
-	// ReadQCQF is the quorum function for the ReadQC
-	// quorum call method. The in parameter is the request object
-	// supplied to the ReadQC method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *ReadRequest'.
-	ReadQCQF(in *ReadRequest, replies map[uint32]*ReadResponse) (*ReadResponse, bool)
-
-	// WriteQCQF is the quorum function for the WriteQC
-	// quorum call method. The in parameter is the request object
-	// supplied to the WriteQC method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *WriteRequest'.
-	WriteQCQF(in *WriteRequest, replies map[uint32]*WriteResponse) (*WriteResponse, bool)
-}
-
 // ReadQC executes the Read Quorum Call on a configuration
 // of Nodes and returns the most recent value.
-func (c *StorageConfiguration) ReadQC(ctx context.Context, in *ReadRequest) (resp *ReadResponse, err error) {
+func (c *StorageConfiguration) ReadQC(ctx context.Context, in *ReadRequest) gorums.Responses[*ReadResponse] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "storage.Storage.ReadQC",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*ReadResponse, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*ReadResponse)
-		}
-		return c.qspec.ReadQCQF(req.(*ReadRequest), r)
+		Message:      in,
+		Method:       "storage.Storage.ReadQC",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*ReadResponse), err
+	return gorums.QuorumCall[*ReadResponse](ctx, c.RawConfiguration, cd)
 }
 
 // WriteQC executes the Write Quorum Call on a configuration
 // of Nodes and returns true if a majority of Nodes were updated.
-func (c *StorageConfiguration) WriteQC(ctx context.Context, in *WriteRequest) (resp *WriteResponse, err error) {
+func (c *StorageConfiguration) WriteQC(ctx context.Context, in *WriteRequest) gorums.Responses[*WriteResponse] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "storage.Storage.WriteQC",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*WriteResponse, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*WriteResponse)
-		}
-		return c.qspec.WriteQCQF(req.(*WriteRequest), r)
+		Message:      in,
+		Method:       "storage.Storage.WriteQC",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*WriteResponse), err
+	return gorums.QuorumCall[*WriteResponse](ctx, c.RawConfiguration, cd)
 }
 
 // ReadRPC executes the Read RPC on a single Node
@@ -295,16 +242,4 @@ func RegisterStorageServer(srv *gorums.Server, impl StorageServer) {
 		defer ctx.Release()
 		impl.WriteMulticast(ctx, req)
 	})
-}
-
-type internalReadResponse struct {
-	nid   uint32
-	reply *ReadResponse
-	err   error
-}
-
-type internalWriteResponse struct {
-	nid   uint32
-	reply *WriteResponse
-	err   error
 }
