@@ -57,33 +57,9 @@ var file_gorums_proto_extTypes = []protoimpl.ExtensionInfo{
 	{
 		ExtendedType:  (*descriptorpb.MethodOptions)(nil),
 		ExtensionType: (*bool)(nil),
-		Field:         50005,
-		Name:          "gorums.correctable",
-		Tag:           "varint,50005,opt,name=correctable",
-		Filename:      "gorums.proto",
-	},
-	{
-		ExtendedType:  (*descriptorpb.MethodOptions)(nil),
-		ExtensionType: (*bool)(nil),
-		Field:         50010,
-		Name:          "gorums.async",
-		Tag:           "varint,50010,opt,name=async",
-		Filename:      "gorums.proto",
-	},
-	{
-		ExtendedType:  (*descriptorpb.MethodOptions)(nil),
-		ExtensionType: (*bool)(nil),
 		Field:         50020,
 		Name:          "gorums.per_node_arg",
 		Tag:           "varint,50020,opt,name=per_node_arg",
-		Filename:      "gorums.proto",
-	},
-	{
-		ExtendedType:  (*descriptorpb.MethodOptions)(nil),
-		ExtensionType: (*string)(nil),
-		Field:         50030,
-		Name:          "gorums.custom_return_type",
-		Tag:           "bytes,50030,opt,name=custom_return_type",
 		Filename:      "gorums.proto",
 	},
 }
@@ -100,16 +76,10 @@ var (
 	E_Multicast = &file_gorums_proto_extTypes[2]
 	// optional bool quorumcall = 50004;
 	E_Quorumcall = &file_gorums_proto_extTypes[3]
-	// optional bool correctable = 50005;
-	E_Correctable = &file_gorums_proto_extTypes[4]
 	// options for call types
 	//
-	// optional bool async = 50010;
-	E_Async = &file_gorums_proto_extTypes[5]
 	// optional bool per_node_arg = 50020;
-	E_PerNodeArg = &file_gorums_proto_extTypes[6]
-	// optional string custom_return_type = 50030;
-	E_CustomReturnType = &file_gorums_proto_extTypes[7]
+	E_PerNodeArg = &file_gorums_proto_extTypes[4]
 )
 
 var File_gorums_proto protoreflect.FileDescriptor
@@ -123,11 +93,8 @@ const file_gorums_proto_rawDesc = "" +
 	"\n" +
 	"quorumcall\x12\x1e.google.protobuf.MethodOptions\x18Ԇ\x03 \x01(\bR\n" +
 	"quorumcall:B\n" +
-	"\vcorrectable\x12\x1e.google.protobuf.MethodOptions\x18Ն\x03 \x01(\bR\vcorrectable:6\n" +
-	"\x05async\x12\x1e.google.protobuf.MethodOptions\x18چ\x03 \x01(\bR\x05async:B\n" +
 	"\fper_node_arg\x12\x1e.google.protobuf.MethodOptions\x18\xe4\x86\x03 \x01(\bR\n" +
-	"perNodeArg:N\n" +
-	"\x12custom_return_type\x12\x1e.google.protobuf.MethodOptions\x18\xee\x86\x03 \x01(\tR\x10customReturnTypeB\x1eZ\x17github.com/relab/gorums\x92\x03\x02\b\x02b\beditionsp\xe8\a"
+	"perNodeArgB\x1eZ\x17github.com/relab/gorums\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
 var file_gorums_proto_goTypes = []any{
 	(*descriptorpb.MethodOptions)(nil), // 0: google.protobuf.MethodOptions
@@ -137,14 +104,11 @@ var file_gorums_proto_depIdxs = []int32{
 	0, // 1: gorums.unicast:extendee -> google.protobuf.MethodOptions
 	0, // 2: gorums.multicast:extendee -> google.protobuf.MethodOptions
 	0, // 3: gorums.quorumcall:extendee -> google.protobuf.MethodOptions
-	0, // 4: gorums.correctable:extendee -> google.protobuf.MethodOptions
-	0, // 5: gorums.async:extendee -> google.protobuf.MethodOptions
-	0, // 6: gorums.per_node_arg:extendee -> google.protobuf.MethodOptions
-	0, // 7: gorums.custom_return_type:extendee -> google.protobuf.MethodOptions
-	8, // [8:8] is the sub-list for method output_type
-	8, // [8:8] is the sub-list for method input_type
-	8, // [8:8] is the sub-list for extension type_name
-	0, // [0:8] is the sub-list for extension extendee
+	0, // 4: gorums.per_node_arg:extendee -> google.protobuf.MethodOptions
+	5, // [5:5] is the sub-list for method output_type
+	5, // [5:5] is the sub-list for method input_type
+	5, // [5:5] is the sub-list for extension type_name
+	0, // [0:5] is the sub-list for extension extendee
 	0, // [0:0] is the sub-list for field type_name
 }
 
@@ -160,7 +124,7 @@ func file_gorums_proto_init() {
 			RawDescriptor: unsafe.Slice(unsafe.StringData(file_gorums_proto_rawDesc), len(file_gorums_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   0,
-			NumExtensions: 8,
+			NumExtensions: 5,
 			NumServices:   0,
 		},
 		GoTypes:           file_gorums_proto_goTypes,
