@@ -3,7 +3,7 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.36.3
+// 	protoc-gen-go v1.36.6
 // 	protoc        v6.30.2
 // source: failing/reservednames/reserved.proto
 
@@ -14,6 +14,7 @@ import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
+	unsafe "unsafe"
 )
 
 const (
@@ -139,40 +140,17 @@ func (b0 Manager_builder) Build() *Manager {
 
 var File_failing_reservednames_reserved_proto protoreflect.FileDescriptor
 
-var file_failing_reservednames_reserved_proto_rawDesc = []byte{
-	0x0a, 0x24, 0x66, 0x61, 0x69, 0x6c, 0x69, 0x6e, 0x67, 0x2f, 0x72, 0x65, 0x73, 0x65, 0x72, 0x76,
-	0x65, 0x64, 0x6e, 0x61, 0x6d, 0x65, 0x73, 0x2f, 0x72, 0x65, 0x73, 0x65, 0x72, 0x76, 0x65, 0x64,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x27, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2e, 0x74,
-	0x65, 0x73, 0x74, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x73, 0x2e, 0x66, 0x61, 0x69, 0x6c, 0x69, 0x6e,
-	0x67, 0x2e, 0x72, 0x65, 0x73, 0x65, 0x72, 0x76, 0x65, 0x64, 0x6e, 0x61, 0x6d, 0x65, 0x73, 0x1a,
-	0x0c, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x23, 0x0a,
-	0x0d, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x12,
-	0x0a, 0x04, 0x43, 0x6f, 0x6e, 0x66, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x43, 0x6f,
-	0x6e, 0x66, 0x22, 0x1b, 0x0a, 0x07, 0x4d, 0x61, 0x6e, 0x61, 0x67, 0x65, 0x72, 0x12, 0x10, 0x0a,
-	0x03, 0x4d, 0x67, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x4d, 0x67, 0x72, 0x32,
-	0x80, 0x02, 0x0a, 0x08, 0x52, 0x65, 0x73, 0x65, 0x72, 0x76, 0x65, 0x64, 0x12, 0x76, 0x0a, 0x08,
-	0x43, 0x6f, 0x6e, 0x66, 0x54, 0x65, 0x73, 0x74, 0x12, 0x36, 0x2e, 0x67, 0x6f, 0x72, 0x75, 0x6d,
-	0x73, 0x2e, 0x74, 0x65, 0x73, 0x74, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x73, 0x2e, 0x66, 0x61, 0x69,
-	0x6c, 0x69, 0x6e, 0x67, 0x2e, 0x72, 0x65, 0x73, 0x65, 0x72, 0x76, 0x65, 0x64, 0x6e, 0x61, 0x6d,
-	0x65, 0x73, 0x2e, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e,
-	0x1a, 0x30, 0x2e, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2e, 0x74, 0x65, 0x73, 0x74, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x73, 0x2e, 0x66, 0x61, 0x69, 0x6c, 0x69, 0x6e, 0x67, 0x2e, 0x72, 0x65, 0x73,
-	0x65, 0x72, 0x76, 0x65, 0x64, 0x6e, 0x61, 0x6d, 0x65, 0x73, 0x2e, 0x4d, 0x61, 0x6e, 0x61, 0x67,
-	0x65, 0x72, 0x22, 0x00, 0x12, 0x7c, 0x0a, 0x0a, 0x51, 0x75, 0x6f, 0x72, 0x75, 0x6d, 0x43, 0x61,
-	0x6c, 0x6c, 0x12, 0x36, 0x2e, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2e, 0x74, 0x65, 0x73, 0x74,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x73, 0x2e, 0x66, 0x61, 0x69, 0x6c, 0x69, 0x6e, 0x67, 0x2e, 0x72,
-	0x65, 0x73, 0x65, 0x72, 0x76, 0x65, 0x64, 0x6e, 0x61, 0x6d, 0x65, 0x73, 0x2e, 0x43, 0x6f, 0x6e,
-	0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x1a, 0x30, 0x2e, 0x67, 0x6f, 0x72,
-	0x75, 0x6d, 0x73, 0x2e, 0x74, 0x65, 0x73, 0x74, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x73, 0x2e, 0x66,
-	0x61, 0x69, 0x6c, 0x69, 0x6e, 0x67, 0x2e, 0x72, 0x65, 0x73, 0x65, 0x72, 0x76, 0x65, 0x64, 0x6e,
-	0x61, 0x6d, 0x65, 0x73, 0x2e, 0x4d, 0x61, 0x6e, 0x61, 0x67, 0x65, 0x72, 0x22, 0x04, 0xa0, 0xb5,
-	0x18, 0x01, 0x42, 0x48, 0x5a, 0x41, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d,
-	0x2f, 0x72, 0x65, 0x6c, 0x61, 0x62, 0x2f, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2f, 0x69, 0x6e,
-	0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x2f, 0x74, 0x65, 0x73, 0x74, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x73, 0x2f, 0x66, 0x61, 0x69, 0x6c, 0x69, 0x6e, 0x67, 0x2f, 0x72, 0x65, 0x73, 0x65, 0x72, 0x76,
-	0x65, 0x64, 0x6e, 0x61, 0x6d, 0x65, 0x73, 0x92, 0x03, 0x02, 0x08, 0x02, 0x62, 0x08, 0x65, 0x64,
-	0x69, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x70, 0xe8, 0x07,
-}
+const file_failing_reservednames_reserved_proto_rawDesc = "" +
+	"\n" +
+	"$failing/reservednames/reserved.proto\x12'gorums.testprotos.failing.reservednames\x1a\fgorums.proto\"#\n" +
+	"\rConfiguration\x12\x12\n" +
+	"\x04Conf\x18\x01 \x01(\tR\x04Conf\"\x1b\n" +
+	"\aManager\x12\x10\n" +
+	"\x03Mgr\x18\x01 \x01(\tR\x03Mgr2\x80\x02\n" +
+	"\bReserved\x12v\n" +
+	"\bConfTest\x126.gorums.testprotos.failing.reservednames.Configuration\x1a0.gorums.testprotos.failing.reservednames.Manager\"\x00\x12|\n" +
+	"\n" +
+	"QuorumCall\x126.gorums.testprotos.failing.reservednames.Configuration\x1a0.gorums.testprotos.failing.reservednames.Manager\"\x04\xa0\xb5\x18\x01BHZAgithub.com/relab/gorums/internal/testprotos/failing/reservednames\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
 var file_failing_reservednames_reserved_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
 var file_failing_reservednames_reserved_proto_goTypes = []any{
@@ -200,7 +178,7 @@ func file_failing_reservednames_reserved_proto_init() {
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_failing_reservednames_reserved_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_failing_reservednames_reserved_proto_rawDesc), len(file_failing_reservednames_reserved_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   2,
 			NumExtensions: 0,
@@ -211,7 +189,6 @@ func file_failing_reservednames_reserved_proto_init() {
 		MessageInfos:      file_failing_reservednames_reserved_proto_msgTypes,
 	}.Build()
 	File_failing_reservednames_reserved_proto = out.File
-	file_failing_reservednames_reserved_proto_rawDesc = nil
 	file_failing_reservednames_reserved_proto_goTypes = nil
 	file_failing_reservednames_reserved_proto_depIdxs = nil
 }
