@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
+// 	protoc-gen-gorums v0.9.0-devel+d97bfa83
 // 	protoc            v6.30.2
 // source: failing/reservednames/reserved.proto
 
@@ -8,17 +8,15 @@ package reservednames
 
 import (
 	context "context"
-	fmt "fmt"
 	gorums "github.com/relab/gorums"
 	encoding "google.golang.org/grpc/encoding"
-	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 func init() {
@@ -29,7 +27,7 @@ func init() {
 
 // ReservedClient is the client interface for the Reserved service.
 type ReservedClient interface {
-	QuorumCall(ctx context.Context, in *Configuration) (resp *Manager, err error)
+	QuorumCall(ctx context.Context, in *Configuration) gorums.Responses[*Manager]
 }
 
 // enforce interface compliance
@@ -47,29 +45,47 @@ var _ ReservedNodeClient = (*ReservedNode)(nil)
 // procedure calls may be invoked.
 type ReservedConfiguration struct {
 	gorums.RawConfiguration
-	qspec ReservedQuorumSpec
-	nodes []*ReservedNode
 }
 
-// ReservedQuorumSpecFromRaw returns a new ReservedQuorumSpec from the given raw configuration and QuorumSpec.
+// NewReservedConfiguration returns a configuration based on the provided list of nodes (required)
+// and an optional quorum specification. The QuorumSpec is necessary for call types that
+// must process replies. For configurations only used for unicast or multicast call types,
+// a QuorumSpec is not needed.
+// Nodes can be supplied using WithNodeMap or WithNodeList.
+// Using any other type of NodeListOption will not work.
+// The ManagerOption list controls how the nodes in the configuration are created.
+func NewReservedConfiguration(cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *ReservedConfiguration, err error) {
+	c = &ReservedConfiguration{}
+	c.RawConfiguration, err = gorums.NewRawConfiguration(cfg, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return c, nil
+}
+
+// SubReservedConfiguration allows for making a new Configuration from the
+// ManagerOption list and node list of another set of configurations,
+// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
+// A new configuration can also be created from an existing configuration,
+// using the And, WithNewNodes, Except, and WithoutNodes methods.
+func (c *ReservedConfiguration) SubReservedConfiguration(cfg gorums.NodeListOption) (subCfg *ReservedConfiguration, err error) {
+	subCfg = &ReservedConfiguration{}
+	subCfg.RawConfiguration, err = c.SubRawConfiguration(cfg)
+	if err != nil {
+		return nil, err
+	}
+	return subCfg, nil
+}
+
+// ReservedConfigurationFromRaw returns a new ReservedConfiguration from the given raw configuration.
 //
 // This function may for example be used to "clone" a configuration but install a different QuorumSpec:
 //
 //	cfg1, err := mgr.NewConfiguration(qspec1, opts...)
 //	cfg2 := ConfigurationFromRaw(cfg1.RawConfig, qspec2)
-func ReservedConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec ReservedQuorumSpec) (*ReservedConfiguration, error) {
-	// return an error if qspec is nil.
-	if qspec == nil {
-		return nil, fmt.Errorf("config: missing required QuorumSpec")
-	}
+func ReservedConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*ReservedConfiguration, error) {
 	newCfg := &ReservedConfiguration{
 		RawConfiguration: rawCfg,
-		qspec:            qspec,
-	}
-	// initialize the nodes slice
-	newCfg.nodes = make([]*ReservedNode, newCfg.Size())
-	for i, n := range rawCfg {
-		newCfg.nodes[i] = &ReservedNode{n}
 	}
 	return newCfg, nil
 }
@@ -79,7 +95,24 @@ func ReservedConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec Reserved
 //
 // NOTE: mutating the returned slice is not supported.
 func (c *ReservedConfiguration) Nodes() []*ReservedNode {
-	return c.nodes
+	rawNodes := c.RawConfiguration.Nodes()
+	nodes := make([]*ReservedNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &ReservedNode{n}
+	}
+	return nodes
+}
+
+// AllNodes returns a slice of each available node of all subconfigurations. Sorted by node id.
+//
+// NOTE: mutating the returned slice is not supported.
+func (c *ReservedConfiguration) AllNodes() []*ReservedNode {
+	rawNodes := c.RawConfiguration.AllNodes()
+	nodes := make([]*ReservedNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &ReservedNode{n}
+	}
+	return nodes
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
@@ -93,96 +126,27 @@ func (c ReservedConfiguration) Except(rm *ReservedConfiguration) gorums.NodeList
 	return c.RawConfiguration.Except(rm.RawConfiguration)
 }
 
-// ReservedManager maintains a connection pool of nodes on
-// which quorum calls can be performed.
-type ReservedManager struct {
-	*gorums.RawManager
-}
-
-// NewReservedManager returns a new ReservedManager for managing connection to nodes added
-// to the manager. This function accepts manager options used to configure
-// various aspects of the manager.
-func NewReservedManager(opts ...gorums.ManagerOption) *ReservedManager {
-	return &ReservedManager{
-		RawManager: gorums.NewRawManager(opts...),
-	}
-}
-
-// NewReservedConfiguration returns a ReservedConfiguration based on the provided list of nodes (required)and a quorum specification.
-// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
-// A new configuration can also be created from an existing configuration,
-// using the And, WithNewNodes, Except, and WithoutNodes methods.
-func (m *ReservedManager) NewConfiguration(cfg gorums.NodeListOption, qspec ReservedQuorumSpec) (c *ReservedConfiguration, err error) {
-	c = &ReservedConfiguration{}
-	c.RawConfiguration, err = gorums.NewRawConfiguration(m.RawManager, cfg)
-	if err != nil {
-		return nil, err
-	}
-	// return an error if qspec is nil.
-	if qspec == nil {
-		return nil, fmt.Errorf("config: missing required ReservedQuorumSpec")
-	}
-	c.qspec = qspec
-	// initialize the nodes slice
-	c.nodes = make([]*ReservedNode, c.Size())
-	for i, n := range c.RawConfiguration {
-		c.nodes[i] = &ReservedNode{n}
-	}
-	return c, nil
-}
-
-// Nodes returns a slice of available nodes on this manager.
-// IDs are returned in the order they were added at creation of the manager.
-func (m *ReservedManager) Nodes() []*ReservedNode {
-	gorumsNodes := m.RawManager.Nodes()
-	nodes := make([]*ReservedNode, len(gorumsNodes))
-	for i, n := range gorumsNodes {
-		nodes[i] = &ReservedNode{n}
-	}
-	return nodes
-}
-
 // ReservedNode holds the node specific methods for the Reserved service.
 type ReservedNode struct {
 	*gorums.RawNode
 }
 
-// ReservedQuorumSpec is the interface of quorum functions for Reserved.
-type ReservedQuorumSpec interface {
-	gorums.ConfigOption
+// QuorumCall is a quorum call invoked on each node in configuration c,
 
-	// QuorumCallQF is the quorum function for the QuorumCall
-	// quorum call method. The in parameter is the request object
-	// supplied to the QuorumCall method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *Configuration'.
-	QuorumCallQF(in *Configuration, replies map[uint32]*Manager) (*Manager, bool)
-}
-
-// QuorumCall is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
-func (c *ReservedConfiguration) QuorumCall(ctx context.Context, in *Configuration) (resp *Manager, err error) {
+// with the same argument in, and returns the responses as an iterator.
+func (c *ReservedConfiguration) QuorumCall(ctx context.Context, in *Configuration) gorums.Responses[*Manager] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "gorums.testprotos.failing.reservednames.Reserved.QuorumCall",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Manager, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Manager)
-		}
-		return c.qspec.QuorumCallQF(req.(*Configuration), r)
+		Message:      in,
+		Method:       "gorums.testprotos.failing.reservednames.Reserved.QuorumCall",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*Manager), err
+	return gorums.QuorumCall[*Manager](ctx, c.RawConfiguration, cd)
 }
 
-// ConfTest is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
+// ConfTest is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
 func (n *ReservedNode) ConfTest(ctx context.Context, in *Configuration) (resp *Manager, err error) {
 	cd := gorums.CallData{
 		Message: in,
@@ -215,10 +179,4 @@ func RegisterReservedServer(srv *gorums.Server, impl ReservedServer) {
 		resp, err := impl.QuorumCall(ctx, req)
 		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
 	})
-}
-
-type internalManager struct {
-	nid   uint32
-	reply *Manager
-	err   error
 }
