@@ -28,6 +28,7 @@ type Metadata struct {
 	xxx_hidden_MessageID uint64                 `protobuf:"varint,1,opt,name=MessageID"`
 	xxx_hidden_Method    string                 `protobuf:"bytes,2,opt,name=Method"`
 	xxx_hidden_Status    *status.Status         `protobuf:"bytes,3,opt,name=Status"`
+	xxx_hidden_Entry     *[]*MetadataEntry      `protobuf:"bytes,4,rep,name=Entry"`
 	unknownFields        protoimpl.UnknownFields
 	sizeCache            protoimpl.SizeCache
 }
@@ -78,6 +79,15 @@ func (x *Metadata) GetStatus() *status.Status {
 	return nil
 }
 
+func (x *Metadata) GetEntry() []*MetadataEntry {
+	if x != nil {
+		if x.xxx_hidden_Entry != nil {
+			return *x.xxx_hidden_Entry
+		}
+	}
+	return nil
+}
+
 func (x *Metadata) SetMessageID(v uint64) {
 	x.xxx_hidden_MessageID = v
 }
@@ -88,6 +98,10 @@ func (x *Metadata) SetMethod(v string) {
 
 func (x *Metadata) SetStatus(v *status.Status) {
 	x.xxx_hidden_Status = v
+}
+
+func (x *Metadata) SetEntry(v []*MetadataEntry) {
+	x.xxx_hidden_Entry = &v
 }
 
 func (x *Metadata) HasStatus() bool {
@@ -107,6 +121,8 @@ type Metadata_builder struct {
 	MessageID uint64
 	Method    string
 	Status    *status.Status
+	// per message client-generated metadata
+	Entry []*MetadataEntry
 }
 
 func (b0 Metadata_builder) Build() *Metadata {
@@ -116,6 +132,78 @@ func (b0 Metadata_builder) Build() *Metadata {
 	x.xxx_hidden_MessageID = b.MessageID
 	x.xxx_hidden_Method = b.Method
 	x.xxx_hidden_Status = b.Status
+	x.xxx_hidden_Entry = &b.Entry
+	return m0
+}
+
+type MetadataEntry struct {
+	state            protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Key   string                 `protobuf:"bytes,1,opt,name=Key"`
+	xxx_hidden_Value string                 `protobuf:"bytes,2,opt,name=Value"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
+}
+
+func (x *MetadataEntry) Reset() {
+	*x = MetadataEntry{}
+	mi := &file_ordering_ordering_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *MetadataEntry) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*MetadataEntry) ProtoMessage() {}
+
+func (x *MetadataEntry) ProtoReflect() protoreflect.Message {
+	mi := &file_ordering_ordering_proto_msgTypes[1]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+func (x *MetadataEntry) GetKey() string {
+	if x != nil {
+		return x.xxx_hidden_Key
+	}
+	return ""
+}
+
+func (x *MetadataEntry) GetValue() string {
+	if x != nil {
+		return x.xxx_hidden_Value
+	}
+	return ""
+}
+
+func (x *MetadataEntry) SetKey(v string) {
+	x.xxx_hidden_Key = v
+}
+
+func (x *MetadataEntry) SetValue(v string) {
+	x.xxx_hidden_Value = v
+}
+
+type MetadataEntry_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Key   string
+	Value string
+}
+
+func (b0 MetadataEntry_builder) Build() *MetadataEntry {
+	m0 := &MetadataEntry{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Key = b.Key
+	x.xxx_hidden_Value = b.Value
 	return m0
 }
 
@@ -123,29 +211,35 @@ var File_ordering_ordering_proto protoreflect.FileDescriptor
 
 const file_ordering_ordering_proto_rawDesc = "" +
 	"\n" +
-	"\x17ordering/ordering.proto\x12\bordering\x1a\x17google/rpc/status.proto\"l\n" +
+	"\x17ordering/ordering.proto\x12\bordering\x1a\x17google/rpc/status.proto\"\x9b\x01\n" +
 	"\bMetadata\x12\x1c\n" +
 	"\tMessageID\x18\x01 \x01(\x04R\tMessageID\x12\x16\n" +
 	"\x06Method\x18\x02 \x01(\tR\x06Method\x12*\n" +
-	"\x06Status\x18\x03 \x01(\v2\x12.google.rpc.StatusR\x06Status2B\n" +
+	"\x06Status\x18\x03 \x01(\v2\x12.google.rpc.StatusR\x06Status\x12-\n" +
+	"\x05Entry\x18\x04 \x03(\v2\x17.ordering.MetadataEntryR\x05Entry\"7\n" +
+	"\rMetadataEntry\x12\x10\n" +
+	"\x03Key\x18\x01 \x01(\tR\x03Key\x12\x14\n" +
+	"\x05Value\x18\x02 \x01(\tR\x05Value2B\n" +
 	"\x06Gorums\x128\n" +
 	"\n" +
 	"NodeStream\x12\x12.ordering.Metadata\x1a\x12.ordering.Metadata(\x010\x01B'Z github.com/relab/gorums/ordering\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
-var file_ordering_ordering_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
+var file_ordering_ordering_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
 var file_ordering_ordering_proto_goTypes = []any{
 	(*Metadata)(nil),      // 0: ordering.Metadata
-	(*status.Status)(nil), // 1: google.rpc.Status
+	(*MetadataEntry)(nil), // 1: ordering.MetadataEntry
+	(*status.Status)(nil), // 2: google.rpc.Status
 }
 var file_ordering_ordering_proto_depIdxs = []int32{
-	1, // 0: ordering.Metadata.Status:type_name -> google.rpc.Status
-	0, // 1: ordering.Gorums.NodeStream:input_type -> ordering.Metadata
-	0, // 2: ordering.Gorums.NodeStream:output_type -> ordering.Metadata
-	2, // [2:3] is the sub-list for method output_type
-	1, // [1:2] is the sub-list for method input_type
-	1, // [1:1] is the sub-list for extension type_name
-	1, // [1:1] is the sub-list for extension extendee
-	0, // [0:1] is the sub-list for field type_name
+	2, // 0: ordering.Metadata.Status:type_name -> google.rpc.Status
+	1, // 1: ordering.Metadata.Entry:type_name -> ordering.MetadataEntry
+	0, // 2: ordering.Gorums.NodeStream:input_type -> ordering.Metadata
+	0, // 3: ordering.Gorums.NodeStream:output_type -> ordering.Metadata
+	3, // [3:4] is the sub-list for method output_type
+	2, // [2:3] is the sub-list for method input_type
+	2, // [2:2] is the sub-list for extension type_name
+	2, // [2:2] is the sub-list for extension extendee
+	0, // [0:2] is the sub-list for field type_name
 }
 
 func init() { file_ordering_ordering_proto_init() }
@@ -159,7 +253,7 @@ func file_ordering_ordering_proto_init() {
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: unsafe.Slice(unsafe.StringData(file_ordering_ordering_proto_rawDesc), len(file_ordering_ordering_proto_rawDesc)),
 			NumEnums:      0,
-			NumMessages:   1,
+			NumMessages:   2,
 			NumExtensions: 0,
 			NumServices:   1,
 		},
