@@ -18,17 +18,15 @@ func TestRPCCallSuccess(t *testing.T) {
 	})
 	defer teardown()
 
-	mgr := gorumsTestMgr()
-
-	_, err := mgr.NewConfiguration(gorums.WithNodeList(addrs))
+	cfg, err := dummy.NewDummyConfiguration(gorums.WithNodeList(addrs), gorumsTestMgrOpts()...)
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	node := mgr.Nodes()[0]
+	node := cfg.Nodes()[0]
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
-	response, err := node.RPCCall(ctx, gorums.CallData{
+	response, err := node.RawNode.RPCCall(ctx, gorums.CallData{
 		Message: &dummy.Empty{},
 		Method:  "dummy.Dummy.Test",
 	})
@@ -44,19 +42,18 @@ func TestRPCCallDownedNode(t *testing.T) {
 	addrs, teardown := gorums.TestSetup(t, 1, func(_ int) gorums.ServerIface {
 		return initServer()
 	})
-	mgr := gorumsTestMgr()
 
-	_, err := mgr.NewConfiguration(gorums.WithNodeList(addrs))
+	cfg, err := dummy.NewDummyConfiguration(gorums.WithNodeList(addrs), gorumsTestMgrOpts()...)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	teardown()                         // stop all servers on purpose
 	time.Sleep(300 * time.Millisecond) // servers are not stopped immediately
-	node := mgr.Nodes()[0]
+	node := cfg.Nodes()[0]
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
-	response, err := node.RPCCall(ctx, gorums.CallData{
+	response, err := node.RawNode.RPCCall(ctx, gorums.CallData{
 		Message: &dummy.Empty{},
 		Method:  "dummy.Dummy.Test",
 	})
@@ -74,18 +71,16 @@ func TestRPCCallTimedOut(t *testing.T) {
 	})
 	defer teardown()
 
-	mgr := gorumsTestMgr()
-
-	_, err := mgr.NewConfiguration(gorums.WithNodeList(addrs))
+	cfg, err := dummy.NewDummyConfiguration(gorums.WithNodeList(addrs), gorumsTestMgrOpts()...)
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	node := mgr.Nodes()[0]
+	node := cfg.Nodes()[0]
 	ctx, cancel := context.WithTimeout(context.Background(), 0*time.Second)
 	time.Sleep(50 * time.Millisecond)
 	defer cancel()
-	response, err := node.RPCCall(ctx, gorums.CallData{
+	response, err := node.RawNode.RPCCall(ctx, gorums.CallData{
 		Message: &dummy.Empty{},
 		Method:  "dummy.Dummy.Test",
 	})
@@ -103,13 +98,11 @@ func initServer() *gorums.Server {
 	return srv
 }
 
-func gorumsTestMgr() *dummy.DummyManager {
-	mgr := dummy.NewDummyManager(
-		gorums.WithGrpcDialOptions(
-			grpc.WithTransportCredentials(insecure.NewCredentials()),
-		),
-	)
-	return mgr
+func gorumsTestMgrOpts() []gorums.ManagerOption {
+	opts := []gorums.ManagerOption{gorums.WithGrpcDialOptions(
+		grpc.WithTransportCredentials(insecure.NewCredentials()),
+	)}
+	return opts
 }
 
 type testSrv struct{}
