@@ -11,7 +11,6 @@ import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
-	sync "sync"
 	unsafe "unsafe"
 )
 
@@ -23,10 +22,10 @@ const (
 )
 
 type Request struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
-	Num           uint64                 `protobuf:"varint,1,opt,name=Num" json:"Num,omitempty"`
-	unknownFields protoimpl.UnknownFields
-	sizeCache     protoimpl.SizeCache
+	state          protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Num uint64                 `protobuf:"varint,1,opt,name=Num"`
+	unknownFields  protoimpl.UnknownFields
+	sizeCache      protoimpl.SizeCache
 }
 
 func (x *Request) Reset() {
@@ -54,24 +53,37 @@ func (x *Request) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Request.ProtoReflect.Descriptor instead.
-func (*Request) Descriptor() ([]byte, []int) {
-	return file_config_config_proto_rawDescGZIP(), []int{0}
-}
-
 func (x *Request) GetNum() uint64 {
 	if x != nil {
-		return x.Num
+		return x.xxx_hidden_Num
 	}
 	return 0
 }
 
+func (x *Request) SetNum(v uint64) {
+	x.xxx_hidden_Num = v
+}
+
+type Request_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Num uint64
+}
+
+func (b0 Request_builder) Build() *Request {
+	m0 := &Request{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Num = b.Num
+	return m0
+}
+
 type Response struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
-	Name          string                 `protobuf:"bytes,1,opt,name=Name" json:"Name,omitempty"`
-	Num           uint64                 `protobuf:"varint,2,opt,name=Num" json:"Num,omitempty"`
-	unknownFields protoimpl.UnknownFields
-	sizeCache     protoimpl.SizeCache
+	state           protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Name string                 `protobuf:"bytes,1,opt,name=Name"`
+	xxx_hidden_Num  uint64                 `protobuf:"varint,2,opt,name=Num"`
+	unknownFields   protoimpl.UnknownFields
+	sizeCache       protoimpl.SizeCache
 }
 
 func (x *Response) Reset() {
@@ -99,23 +111,42 @@ func (x *Response) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Response.ProtoReflect.Descriptor instead.
-func (*Response) Descriptor() ([]byte, []int) {
-	return file_config_config_proto_rawDescGZIP(), []int{1}
-}
-
 func (x *Response) GetName() string {
 	if x != nil {
-		return x.Name
+		return x.xxx_hidden_Name
 	}
 	return ""
 }
 
 func (x *Response) GetNum() uint64 {
 	if x != nil {
-		return x.Num
+		return x.xxx_hidden_Num
 	}
 	return 0
+}
+
+func (x *Response) SetName(v string) {
+	x.xxx_hidden_Name = v
+}
+
+func (x *Response) SetNum(v uint64) {
+	x.xxx_hidden_Num = v
+}
+
+type Response_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Name string
+	Num  uint64
+}
+
+func (b0 Response_builder) Build() *Response {
+	m0 := &Response{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Name = b.Name
+	x.xxx_hidden_Num = b.Num
+	return m0
 }
 
 var File_config_config_proto protoreflect.FileDescriptor
@@ -131,18 +162,6 @@ const file_config_config_proto_rawDesc = "" +
 	"\n" +
 	"ConfigTest\x121\n" +
 	"\x06Config\x12\x0f.config.Request\x1a\x10.config.Response\"\x04\xa0\xb5\x18\x01B+Z$github.com/relab/gorums/tests/config\x92\x03\x02\b\x02b\beditionsp\xe8\a"
-
-var (
-	file_config_config_proto_rawDescOnce sync.Once
-	file_config_config_proto_rawDescData []byte
-)
-
-func file_config_config_proto_rawDescGZIP() []byte {
-	file_config_config_proto_rawDescOnce.Do(func() {
-		file_config_config_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_config_config_proto_rawDesc), len(file_config_config_proto_rawDesc)))
-	})
-	return file_config_config_proto_rawDescData
-}
 
 var file_config_config_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
 var file_config_config_proto_goTypes = []any{
