@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
+// 	protoc-gen-gorums v0.9.0-devel+d97bfa83
 // 	protoc            v6.30.2
 // source: config/config.proto
 
@@ -8,17 +8,15 @@ package config
 
 import (
 	context "context"
-	fmt "fmt"
 	gorums "github.com/relab/gorums"
 	encoding "google.golang.org/grpc/encoding"
-	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 func init() {
@@ -29,7 +27,7 @@ func init() {
 
 // ConfigTestClient is the client interface for the ConfigTest service.
 type ConfigTestClient interface {
-	Config(ctx context.Context, in *Request) (resp *Response, err error)
+	Config(ctx context.Context, in *Request) gorums.Responses[*Response]
 }
 
 // enforce interface compliance
@@ -46,29 +44,47 @@ var _ ConfigTestNodeClient = (*ConfigTestNode)(nil)
 // procedure calls may be invoked.
 type ConfigTestConfiguration struct {
 	gorums.RawConfiguration
-	qspec ConfigTestQuorumSpec
-	nodes []*ConfigTestNode
 }
 
-// ConfigTestQuorumSpecFromRaw returns a new ConfigTestQuorumSpec from the given raw configuration and QuorumSpec.
+// NewConfigTestConfiguration returns a configuration based on the provided list of nodes (required)
+// and an optional quorum specification. The QuorumSpec is necessary for call types that
+// must process replies. For configurations only used for unicast or multicast call types,
+// a QuorumSpec is not needed.
+// Nodes can be supplied using WithNodeMap or WithNodeList.
+// Using any other type of NodeListOption will not work.
+// The ManagerOption list controls how the nodes in the configuration are created.
+func NewConfigTestConfiguration(cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *ConfigTestConfiguration, err error) {
+	c = &ConfigTestConfiguration{}
+	c.RawConfiguration, err = gorums.NewRawConfiguration(cfg, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return c, nil
+}
+
+// SubConfigTestConfiguration allows for making a new Configuration from the
+// ManagerOption list and node list of another set of configurations,
+// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
+// A new configuration can also be created from an existing configuration,
+// using the And, WithNewNodes, Except, and WithoutNodes methods.
+func (c *ConfigTestConfiguration) SubConfigTestConfiguration(cfg gorums.NodeListOption) (subCfg *ConfigTestConfiguration, err error) {
+	subCfg = &ConfigTestConfiguration{}
+	subCfg.RawConfiguration, err = c.SubRawConfiguration(cfg)
+	if err != nil {
+		return nil, err
+	}
+	return subCfg, nil
+}
+
+// ConfigTestConfigurationFromRaw returns a new ConfigTestConfiguration from the given raw configuration.
 //
 // This function may for example be used to "clone" a configuration but install a different QuorumSpec:
 //
 //	cfg1, err := mgr.NewConfiguration(qspec1, opts...)
 //	cfg2 := ConfigurationFromRaw(cfg1.RawConfig, qspec2)
-func ConfigTestConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec ConfigTestQuorumSpec) (*ConfigTestConfiguration, error) {
-	// return an error if qspec is nil.
-	if qspec == nil {
-		return nil, fmt.Errorf("config: missing required QuorumSpec")
-	}
+func ConfigTestConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*ConfigTestConfiguration, error) {
 	newCfg := &ConfigTestConfiguration{
 		RawConfiguration: rawCfg,
-		qspec:            qspec,
-	}
-	// initialize the nodes slice
-	newCfg.nodes = make([]*ConfigTestNode, newCfg.Size())
-	for i, n := range rawCfg {
-		newCfg.nodes[i] = &ConfigTestNode{n}
 	}
 	return newCfg, nil
 }
@@ -78,7 +94,24 @@ func ConfigTestConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec Config
 //
 // NOTE: mutating the returned slice is not supported.
 func (c *ConfigTestConfiguration) Nodes() []*ConfigTestNode {
-	return c.nodes
+	rawNodes := c.RawConfiguration.Nodes()
+	nodes := make([]*ConfigTestNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &ConfigTestNode{n}
+	}
+	return nodes
+}
+
+// AllNodes returns a slice of each available node of all subconfigurations. Sorted by node id.
+//
+// NOTE: mutating the returned slice is not supported.
+func (c *ConfigTestConfiguration) AllNodes() []*ConfigTestNode {
+	rawNodes := c.RawConfiguration.AllNodes()
+	nodes := make([]*ConfigTestNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &ConfigTestNode{n}
+	}
+	return nodes
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
@@ -92,92 +125,22 @@ func (c ConfigTestConfiguration) Except(rm *ConfigTestConfiguration) gorums.Node
 	return c.RawConfiguration.Except(rm.RawConfiguration)
 }
 
-// ConfigTestManager maintains a connection pool of nodes on
-// which quorum calls can be performed.
-type ConfigTestManager struct {
-	*gorums.RawManager
-}
-
-// NewConfigTestManager returns a new ConfigTestManager for managing connection to nodes added
-// to the manager. This function accepts manager options used to configure
-// various aspects of the manager.
-func NewConfigTestManager(opts ...gorums.ManagerOption) *ConfigTestManager {
-	return &ConfigTestManager{
-		RawManager: gorums.NewRawManager(opts...),
-	}
-}
-
-// NewConfigTestConfiguration returns a ConfigTestConfiguration based on the provided list of nodes (required)and a quorum specification.
-// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
-// A new configuration can also be created from an existing configuration,
-// using the And, WithNewNodes, Except, and WithoutNodes methods.
-func (m *ConfigTestManager) NewConfiguration(cfg gorums.NodeListOption, qspec ConfigTestQuorumSpec) (c *ConfigTestConfiguration, err error) {
-	c = &ConfigTestConfiguration{}
-	c.RawConfiguration, err = gorums.NewRawConfiguration(m.RawManager, cfg)
-	if err != nil {
-		return nil, err
-	}
-	// return an error if qspec is nil.
-	if qspec == nil {
-		return nil, fmt.Errorf("config: missing required ConfigTestQuorumSpec")
-	}
-	c.qspec = qspec
-	// initialize the nodes slice
-	c.nodes = make([]*ConfigTestNode, c.Size())
-	for i, n := range c.RawConfiguration {
-		c.nodes[i] = &ConfigTestNode{n}
-	}
-	return c, nil
-}
-
-// Nodes returns a slice of available nodes on this manager.
-// IDs are returned in the order they were added at creation of the manager.
-func (m *ConfigTestManager) Nodes() []*ConfigTestNode {
-	gorumsNodes := m.RawManager.Nodes()
-	nodes := make([]*ConfigTestNode, len(gorumsNodes))
-	for i, n := range gorumsNodes {
-		nodes[i] = &ConfigTestNode{n}
-	}
-	return nodes
-}
-
 // ConfigTestNode holds the node specific methods for the ConfigTest service.
 type ConfigTestNode struct {
 	*gorums.RawNode
 }
 
-// ConfigTestQuorumSpec is the interface of quorum functions for ConfigTest.
-type ConfigTestQuorumSpec interface {
-	gorums.ConfigOption
+// Config is a quorum call invoked on each node in configuration c,
 
-	// ConfigQF is the quorum function for the Config
-	// quorum call method. The in parameter is the request object
-	// supplied to the Config method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *Request'.
-	ConfigQF(in *Request, replies map[uint32]*Response) (*Response, bool)
-}
-
-// Config is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
-func (c *ConfigTestConfiguration) Config(ctx context.Context, in *Request) (resp *Response, err error) {
+// with the same argument in, and returns the responses as an iterator.
+func (c *ConfigTestConfiguration) Config(ctx context.Context, in *Request) gorums.Responses[*Response] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "config.ConfigTest.Config",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Response, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Response)
-		}
-		return c.qspec.ConfigQF(req.(*Request), r)
+		Message:      in,
+		Method:       "config.ConfigTest.Config",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*Response), err
+	return gorums.QuorumCall[*Response](ctx, c.RawConfiguration, cd)
 }
 
 // ConfigTest is the server-side API for the ConfigTest Service
@@ -192,10 +155,4 @@ func RegisterConfigTestServer(srv *gorums.Server, impl ConfigTestServer) {
 		resp, err := impl.Config(ctx, req)
 		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
 	})
-}
-
-type internalResponse struct {
-	nid   uint32
-	reply *Response
-	err   error
 }
