@@ -10,7 +10,6 @@ import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
-	sync "sync"
 	unsafe "unsafe"
 )
 
@@ -22,7 +21,7 @@ const (
 )
 
 type Empty struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
+	state         protoimpl.MessageState `protogen:"opaque.v1"`
 	unknownFields protoimpl.UnknownFields
 	sizeCache     protoimpl.SizeCache
 }
@@ -52,9 +51,16 @@ func (x *Empty) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Empty.ProtoReflect.Descriptor instead.
-func (*Empty) Descriptor() ([]byte, []int) {
-	return file_dummy_dummy_proto_rawDescGZIP(), []int{0}
+type Empty_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+}
+
+func (b0 Empty_builder) Build() *Empty {
+	m0 := &Empty{}
+	b, x := &b0, m0
+	_, _ = b, x
+	return m0
 }
 
 var File_dummy_dummy_proto protoreflect.FileDescriptor
@@ -65,18 +71,6 @@ const file_dummy_dummy_proto_rawDesc = "" +
 	"\x05Empty2-\n" +
 	"\x05Dummy\x12$\n" +
 	"\x04Test\x12\f.dummy.Empty\x1a\f.dummy.Empty\"\x00B*Z#github.com/relab/gorums/tests/dummy\x92\x03\x02\b\x02b\beditionsp\xe8\a"
-
-var (
-	file_dummy_dummy_proto_rawDescOnce sync.Once
-	file_dummy_dummy_proto_rawDescData []byte
-)
-
-func file_dummy_dummy_proto_rawDescGZIP() []byte {
-	file_dummy_dummy_proto_rawDescOnce.Do(func() {
-		file_dummy_dummy_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_dummy_dummy_proto_rawDesc), len(file_dummy_dummy_proto_rawDesc)))
-	})
-	return file_dummy_dummy_proto_rawDescData
-}
 
 var file_dummy_dummy_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
 var file_dummy_dummy_proto_goTypes = []any{
