@@ -11,7 +11,6 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	emptypb "google.golang.org/protobuf/types/known/emptypb"
 	reflect "reflect"
-	sync "sync"
 	unsafe "unsafe"
 )
 
@@ -23,8 +22,8 @@ const (
 )
 
 type NodeID struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
-	ID            uint32                 `protobuf:"varint,1,opt,name=ID" json:"ID,omitempty"`
+	state         protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_ID uint32                 `protobuf:"varint,1,opt,name=ID"`
 	unknownFields protoimpl.UnknownFields
 	sizeCache     protoimpl.SizeCache
 }
@@ -54,23 +53,36 @@ func (x *NodeID) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use NodeID.ProtoReflect.Descriptor instead.
-func (*NodeID) Descriptor() ([]byte, []int) {
-	return file_metadata_metadata_proto_rawDescGZIP(), []int{0}
-}
-
 func (x *NodeID) GetID() uint32 {
 	if x != nil {
-		return x.ID
+		return x.xxx_hidden_ID
 	}
 	return 0
 }
 
+func (x *NodeID) SetID(v uint32) {
+	x.xxx_hidden_ID = v
+}
+
+type NodeID_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	ID uint32
+}
+
+func (b0 NodeID_builder) Build() *NodeID {
+	m0 := &NodeID{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_ID = b.ID
+	return m0
+}
+
 type IPAddr struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
-	Addr          string                 `protobuf:"bytes,1,opt,name=Addr" json:"Addr,omitempty"`
-	unknownFields protoimpl.UnknownFields
-	sizeCache     protoimpl.SizeCache
+	state           protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Addr string                 `protobuf:"bytes,1,opt,name=Addr"`
+	unknownFields   protoimpl.UnknownFields
+	sizeCache       protoimpl.SizeCache
 }
 
 func (x *IPAddr) Reset() {
@@ -98,16 +110,29 @@ func (x *IPAddr) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use IPAddr.ProtoReflect.Descriptor instead.
-func (*IPAddr) Descriptor() ([]byte, []int) {
-	return file_metadata_metadata_proto_rawDescGZIP(), []int{1}
-}
-
 func (x *IPAddr) GetAddr() string {
 	if x != nil {
-		return x.Addr
+		return x.xxx_hidden_Addr
 	}
 	return ""
+}
+
+func (x *IPAddr) SetAddr(v string) {
+	x.xxx_hidden_Addr = v
+}
+
+type IPAddr_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Addr string
+}
+
+func (b0 IPAddr_builder) Build() *IPAddr {
+	m0 := &IPAddr{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Addr = b.Addr
+	return m0
 }
 
 var File_metadata_metadata_proto protoreflect.FileDescriptor
@@ -122,18 +147,6 @@ const file_metadata_metadata_proto_rawDesc = "" +
 	"\fMetadataTest\x126\n" +
 	"\bIDFromMD\x12\x16.google.protobuf.Empty\x1a\x10.metadata.NodeID\"\x00\x124\n" +
 	"\x06WhatIP\x12\x16.google.protobuf.Empty\x1a\x10.metadata.IPAddr\"\x00B-Z&github.com/relab/gorums/tests/metadata\x92\x03\x02\b\x02b\beditionsp\xe8\a"
-
-var (
-	file_metadata_metadata_proto_rawDescOnce sync.Once
-	file_metadata_metadata_proto_rawDescData []byte
-)
-
-func file_metadata_metadata_proto_rawDescGZIP() []byte {
-	file_metadata_metadata_proto_rawDescOnce.Do(func() {
-		file_metadata_metadata_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_metadata_metadata_proto_rawDesc), len(file_metadata_metadata_proto_rawDesc)))
-	})
-	return file_metadata_metadata_proto_rawDescData
-}
 
 var file_metadata_metadata_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
 var file_metadata_metadata_proto_goTypes = []any{
