@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
+// 	protoc-gen-gorums v0.9.0-devel+d97bfa83
 // 	protoc            v6.30.2
 // source: metadata/metadata.proto
 
@@ -15,9 +15,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 func init() {
@@ -30,10 +30,39 @@ func init() {
 // procedure calls may be invoked.
 type MetadataTestConfiguration struct {
 	gorums.RawConfiguration
-	nodes []*MetadataTestNode
 }
 
-// MetadataTestQuorumSpecFromRaw returns a new MetadataTestQuorumSpec from the given raw configuration.
+// NewMetadataTestConfiguration returns a configuration based on the provided list of nodes (required)
+// and an optional quorum specification. The QuorumSpec is necessary for call types that
+// must process replies. For configurations only used for unicast or multicast call types,
+// a QuorumSpec is not needed.
+// Nodes can be supplied using WithNodeMap or WithNodeList.
+// Using any other type of NodeListOption will not work.
+// The ManagerOption list controls how the nodes in the configuration are created.
+func NewMetadataTestConfiguration(cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *MetadataTestConfiguration, err error) {
+	c = &MetadataTestConfiguration{}
+	c.RawConfiguration, err = gorums.NewRawConfiguration(cfg, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return c, nil
+}
+
+// SubMetadataTestConfiguration allows for making a new Configuration from the
+// ManagerOption list and node list of another set of configurations,
+// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
+// A new configuration can also be created from an existing configuration,
+// using the And, WithNewNodes, Except, and WithoutNodes methods.
+func (c *MetadataTestConfiguration) SubMetadataTestConfiguration(cfg gorums.NodeListOption) (subCfg *MetadataTestConfiguration, err error) {
+	subCfg = &MetadataTestConfiguration{}
+	subCfg.RawConfiguration, err = c.SubRawConfiguration(cfg)
+	if err != nil {
+		return nil, err
+	}
+	return subCfg, nil
+}
+
+// MetadataTestConfigurationFromRaw returns a new MetadataTestConfiguration from the given raw configuration.
 //
 // This function may for example be used to "clone" a configuration but install a different QuorumSpec:
 //
@@ -43,11 +72,6 @@ func MetadataTestConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*Metadata
 	newCfg := &MetadataTestConfiguration{
 		RawConfiguration: rawCfg,
 	}
-	// initialize the nodes slice
-	newCfg.nodes = make([]*MetadataTestNode, newCfg.Size())
-	for i, n := range rawCfg {
-		newCfg.nodes[i] = &MetadataTestNode{n}
-	}
 	return newCfg, nil
 }
 
@@ -56,7 +80,24 @@ func MetadataTestConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*Metadata
 //
 // NOTE: mutating the returned slice is not supported.
 func (c *MetadataTestConfiguration) Nodes() []*MetadataTestNode {
-	return c.nodes
+	rawNodes := c.RawConfiguration.Nodes()
+	nodes := make([]*MetadataTestNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &MetadataTestNode{n}
+	}
+	return nodes
+}
+
+// AllNodes returns a slice of each available node of all subconfigurations. Sorted by node id.
+//
+// NOTE: mutating the returned slice is not supported.
+func (c *MetadataTestConfiguration) AllNodes() []*MetadataTestNode {
+	rawNodes := c.RawConfiguration.AllNodes()
+	nodes := make([]*MetadataTestNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &MetadataTestNode{n}
+	}
+	return nodes
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
@@ -68,50 +109,6 @@ func (c MetadataTestConfiguration) And(d *MetadataTestConfiguration) gorums.Node
 // from c without the nodes in rm.
 func (c MetadataTestConfiguration) Except(rm *MetadataTestConfiguration) gorums.NodeListOption {
 	return c.RawConfiguration.Except(rm.RawConfiguration)
-}
-
-// MetadataTestManager maintains a connection pool of nodes on
-// which quorum calls can be performed.
-type MetadataTestManager struct {
-	*gorums.RawManager
-}
-
-// NewMetadataTestManager returns a new MetadataTestManager for managing connection to nodes added
-// to the manager. This function accepts manager options used to configure
-// various aspects of the manager.
-func NewMetadataTestManager(opts ...gorums.ManagerOption) *MetadataTestManager {
-	return &MetadataTestManager{
-		RawManager: gorums.NewRawManager(opts...),
-	}
-}
-
-// NewMetadataTestConfiguration returns a MetadataTestConfiguration based on the provided list of nodes (required).
-// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
-// A new configuration can also be created from an existing configuration,
-// using the And, WithNewNodes, Except, and WithoutNodes methods.
-func (m *MetadataTestManager) NewConfiguration(cfg gorums.NodeListOption) (c *MetadataTestConfiguration, err error) {
-	c = &MetadataTestConfiguration{}
-	c.RawConfiguration, err = gorums.NewRawConfiguration(m.RawManager, cfg)
-	if err != nil {
-		return nil, err
-	}
-	// initialize the nodes slice
-	c.nodes = make([]*MetadataTestNode, c.Size())
-	for i, n := range c.RawConfiguration {
-		c.nodes[i] = &MetadataTestNode{n}
-	}
-	return c, nil
-}
-
-// Nodes returns a slice of available nodes on this manager.
-// IDs are returned in the order they were added at creation of the manager.
-func (m *MetadataTestManager) Nodes() []*MetadataTestNode {
-	gorumsNodes := m.RawManager.Nodes()
-	nodes := make([]*MetadataTestNode, len(gorumsNodes))
-	for i, n := range gorumsNodes {
-		nodes[i] = &MetadataTestNode{n}
-	}
-	return nodes
 }
 
 // MetadataTestNode holds the node specific methods for the MetadataTest service.
