@@ -21,17 +21,17 @@ type testSrv struct{}
 func (srv testSrv) IDFromMD(ctx gorums.ServerCtx, _ *emptypb.Empty) (resp *NodeID, err error) {
 	md, ok := metadata.FromIncomingContext(ctx)
 	if !ok {
-		return nil, status.Error(codes.NotFound, "Metadata unavailable")
+		return nil, status.Error(codes.NotFound, "metadata unavailable")
 	}
 	v := md.Get("id")
 	if len(v) < 1 {
-		return nil, status.Error(codes.NotFound, "ID field missing")
+		return nil, status.Error(codes.NotFound, "missing metadata field: id")
 	}
-	id, err := strconv.Atoi(v[0])
+	id, err := strconv.ParseUint(v[0], 10, 32)
 	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, "Got '%s', but could not convert to integer", v[0])
+		return nil, status.Errorf(codes.InvalidArgument, "value of id field: %q is not a number: %v", v[0], err)
 	}
-	return &NodeID{ID: uint32(id)}, nil
+	return NodeID_builder{ID: uint32(id)}.Build(), nil
 }
 
 func (srv testSrv) WhatIP(ctx gorums.ServerCtx, _ *emptypb.Empty) (resp *IPAddr, err error) {
@@ -39,10 +39,10 @@ func (srv testSrv) WhatIP(ctx gorums.ServerCtx, _ *emptypb.Empty) (resp *IPAddr,
 	if !ok {
 		return nil, status.Error(codes.NotFound, "Peer info unavailable")
 	}
-	return &IPAddr{Addr: peerInfo.Addr.String()}, nil
+	return IPAddr_builder{Addr: peerInfo.Addr.String()}.Build(), nil
 }
 
-func initServer(t *testing.T) *gorums.Server {
+func initServer() *gorums.Server {
 	srv := gorums.NewServer()
 	RegisterMetadataTestServer(srv, &testSrv{})
 	return srv
@@ -51,25 +51,25 @@ func initServer(t *testing.T) *gorums.Server {
 func TestMetadata(t *testing.T) {
 	want := uint32(1)
 
-	addrs, teardown := gorums.TestSetup(t, 1, func(_ int) gorums.ServerIface { return initServer(t) })
+	addrs, teardown := gorums.TestSetup(t, 1, func(_ int) gorums.ServerIface { return initServer() })
 	defer teardown()
 
 	md := metadata.New(map[string]string{
 		"id": fmt.Sprint(want),
 	})
 
-	mgr := NewMetadataTestManager(
+	cfg, err := NewMetadataTestConfiguration(
+		gorums.WithNodeList(addrs),
 		gorums.WithMetadata(md),
 		gorums.WithGrpcDialOptions(
 			grpc.WithTransportCredentials(insecure.NewCredentials()),
 		),
 	)
-	_, err := mgr.NewConfiguration(gorums.WithNodeList(addrs))
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	node := mgr.Nodes()[0]
+	node := cfg.Nodes()[0]
 	resp, err := node.IDFromMD(context.Background(), &emptypb.Empty{})
 	if err != nil {
 		t.Fatalf("RPC error: %v", err)
@@ -80,8 +80,39 @@ func TestMetadata(t *testing.T) {
 	}
 }
 
+func TestPerMessageMetadata(t *testing.T) {
+	want := uint32(1)
+
+	addrs, teardown := gorums.TestSetup(t, 1, func(_ int) gorums.ServerIface { return initServer() })
+	defer teardown()
+
+	cfg, err := NewMetadataTestConfiguration(gorums.WithNodeList(addrs),
+		gorums.WithGrpcDialOptions(
+			grpc.WithTransportCredentials(insecure.NewCredentials()),
+		),
+	)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	node := cfg.Nodes()[0]
+
+	md := metadata.New(map[string]string{
+		"id": fmt.Sprint(want),
+	})
+	ctx := metadata.NewOutgoingContext(context.Background(), md)
+	resp, err := node.IDFromMD(ctx, &emptypb.Empty{})
+	if err != nil {
+		t.Fatalf("RPC error: %v", err)
+	}
+
+	if resp.GetID() != want {
+		t.Fatalf("IDFromMD() == %d, want %d", resp.GetID(), want)
+	}
+}
+
 func TestPerNodeMetadata(t *testing.T) {
-	addrs, teardown := gorums.TestSetup(t, 2, func(_ int) gorums.ServerIface { return initServer(t) })
+	addrs, teardown := gorums.TestSetup(t, 2, func(_ int) gorums.ServerIface { return initServer() })
 	defer teardown()
 
 	perNodeMD := func(nid uint32) metadata.MD {
@@ -90,18 +121,18 @@ func TestPerNodeMetadata(t *testing.T) {
 		})
 	}
 
-	mgr := NewMetadataTestManager(
+	cfg, err := NewMetadataTestConfiguration(
+		gorums.WithNodeList(addrs),
 		gorums.WithPerNodeMetadata(perNodeMD),
 		gorums.WithGrpcDialOptions(
 			grpc.WithTransportCredentials(insecure.NewCredentials()),
 		),
 	)
-	_, err := mgr.NewConfiguration(gorums.WithNodeList(addrs))
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	for _, node := range mgr.Nodes() {
+	for _, node := range cfg.Nodes() {
 		resp, err := node.IDFromMD(context.Background(), &emptypb.Empty{})
 		if err != nil {
 			t.Fatalf("RPC error: %v", err)
@@ -114,20 +145,20 @@ func TestPerNodeMetadata(t *testing.T) {
 }
 
 func TestCanGetPeerInfo(t *testing.T) {
-	addrs, teardown := gorums.TestSetup(t, 1, func(_ int) gorums.ServerIface { return initServer(t) })
+	addrs, teardown := gorums.TestSetup(t, 1, func(_ int) gorums.ServerIface { return initServer() })
 	defer teardown()
 
-	mgr := NewMetadataTestManager(
+	cfg, err := NewMetadataTestConfiguration(
+		gorums.WithNodeList(addrs),
 		gorums.WithGrpcDialOptions(
 			grpc.WithTransportCredentials(insecure.NewCredentials()),
 		),
 	)
-	_, err := mgr.NewConfiguration(gorums.WithNodeList(addrs))
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	node := mgr.Nodes()[0]
+	node := cfg.Nodes()[0]
 	ip, err := node.WhatIP(context.Background(), &emptypb.Empty{})
 	if err != nil {
 		t.Fatalf("RPC error: %v", err)
