@@ -11,7 +11,6 @@ import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
-	sync "sync"
 	unsafe "unsafe"
 )
 
@@ -23,10 +22,10 @@ const (
 )
 
 type Request struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
-	Num           uint64                 `protobuf:"varint,1,opt,name=Num" json:"Num,omitempty"`
-	unknownFields protoimpl.UnknownFields
-	sizeCache     protoimpl.SizeCache
+	state          protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Num uint64                 `protobuf:"varint,1,opt,name=Num"`
+	unknownFields  protoimpl.UnknownFields
+	sizeCache      protoimpl.SizeCache
 }
 
 func (x *Request) Reset() {
@@ -54,20 +53,33 @@ func (x *Request) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Request.ProtoReflect.Descriptor instead.
-func (*Request) Descriptor() ([]byte, []int) {
-	return file_oneway_oneway_proto_rawDescGZIP(), []int{0}
-}
-
 func (x *Request) GetNum() uint64 {
 	if x != nil {
-		return x.Num
+		return x.xxx_hidden_Num
 	}
 	return 0
 }
 
+func (x *Request) SetNum(v uint64) {
+	x.xxx_hidden_Num = v
+}
+
+type Request_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Num uint64
+}
+
+func (b0 Request_builder) Build() *Request {
+	m0 := &Request{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Num = b.Num
+	return m0
+}
+
 type Empty struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
+	state         protoimpl.MessageState `protogen:"opaque.v1"`
 	unknownFields protoimpl.UnknownFields
 	sizeCache     protoimpl.SizeCache
 }
@@ -97,9 +109,16 @@ func (x *Empty) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Empty.ProtoReflect.Descriptor instead.
-func (*Empty) Descriptor() ([]byte, []int) {
-	return file_oneway_oneway_proto_rawDescGZIP(), []int{1}
+type Empty_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+}
+
+func (b0 Empty_builder) Build() *Empty {
+	m0 := &Empty{}
+	b, x := &b0, m0
+	_, _ = b, x
+	return m0
 }
 
 var File_oneway_oneway_proto protoreflect.FileDescriptor
@@ -115,18 +134,6 @@ const file_oneway_oneway_proto_rawDesc = "" +
 	"\aUnicast\x12\x0f.oneway.Request\x1a\r.oneway.Empty\"\x04\x90\xb5\x18\x01\x121\n" +
 	"\tMulticast\x12\x0f.oneway.Request\x1a\r.oneway.Empty\"\x04\x98\xb5\x18\x01\x12<\n" +
 	"\x10MulticastPerNode\x12\x0f.oneway.Request\x1a\r.oneway.Empty\"\b\x98\xb5\x18\x01\xa0\xb6\x18\x01B+Z$github.com/relab/gorums/tests/oneway\x92\x03\x02\b\x02b\beditionsp\xe8\a"
-
-var (
-	file_oneway_oneway_proto_rawDescOnce sync.Once
-	file_oneway_oneway_proto_rawDescData []byte
-)
-
-func file_oneway_oneway_proto_rawDescGZIP() []byte {
-	file_oneway_oneway_proto_rawDescOnce.Do(func() {
-		file_oneway_oneway_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_oneway_oneway_proto_rawDesc), len(file_oneway_oneway_proto_rawDesc)))
-	})
-	return file_oneway_oneway_proto_rawDescData
-}
 
 var file_oneway_oneway_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
 var file_oneway_oneway_proto_goTypes = []any{
