@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
+// 	protoc-gen-gorums v0.9.0-devel+d97bfa83
 // 	protoc            v6.30.2
 // source: oneway/oneway.proto
 
@@ -10,14 +10,14 @@ import (
 	context "context"
 	gorums "github.com/relab/gorums"
 	encoding "google.golang.org/grpc/encoding"
-	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
+	proto "google.golang.org/protobuf/proto"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 func init() {
@@ -47,10 +47,39 @@ var _ OnewayTestNodeClient = (*OnewayTestNode)(nil)
 // procedure calls may be invoked.
 type OnewayTestConfiguration struct {
 	gorums.RawConfiguration
-	nodes []*OnewayTestNode
 }
 
-// OnewayTestQuorumSpecFromRaw returns a new OnewayTestQuorumSpec from the given raw configuration.
+// NewOnewayTestConfiguration returns a configuration based on the provided list of nodes (required)
+// and an optional quorum specification. The QuorumSpec is necessary for call types that
+// must process replies. For configurations only used for unicast or multicast call types,
+// a QuorumSpec is not needed.
+// Nodes can be supplied using WithNodeMap or WithNodeList.
+// Using any other type of NodeListOption will not work.
+// The ManagerOption list controls how the nodes in the configuration are created.
+func NewOnewayTestConfiguration(cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *OnewayTestConfiguration, err error) {
+	c = &OnewayTestConfiguration{}
+	c.RawConfiguration, err = gorums.NewRawConfiguration(cfg, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return c, nil
+}
+
+// SubOnewayTestConfiguration allows for making a new Configuration from the
+// ManagerOption list and node list of another set of configurations,
+// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
+// A new configuration can also be created from an existing configuration,
+// using the And, WithNewNodes, Except, and WithoutNodes methods.
+func (c *OnewayTestConfiguration) SubOnewayTestConfiguration(cfg gorums.NodeListOption) (subCfg *OnewayTestConfiguration, err error) {
+	subCfg = &OnewayTestConfiguration{}
+	subCfg.RawConfiguration, err = c.SubRawConfiguration(cfg)
+	if err != nil {
+		return nil, err
+	}
+	return subCfg, nil
+}
+
+// OnewayTestConfigurationFromRaw returns a new OnewayTestConfiguration from the given raw configuration.
 //
 // This function may for example be used to "clone" a configuration but install a different QuorumSpec:
 //
@@ -60,11 +89,6 @@ func OnewayTestConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*OnewayTest
 	newCfg := &OnewayTestConfiguration{
 		RawConfiguration: rawCfg,
 	}
-	// initialize the nodes slice
-	newCfg.nodes = make([]*OnewayTestNode, newCfg.Size())
-	for i, n := range rawCfg {
-		newCfg.nodes[i] = &OnewayTestNode{n}
-	}
 	return newCfg, nil
 }
 
@@ -73,7 +97,24 @@ func OnewayTestConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*OnewayTest
 //
 // NOTE: mutating the returned slice is not supported.
 func (c *OnewayTestConfiguration) Nodes() []*OnewayTestNode {
-	return c.nodes
+	rawNodes := c.RawConfiguration.Nodes()
+	nodes := make([]*OnewayTestNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &OnewayTestNode{n}
+	}
+	return nodes
+}
+
+// AllNodes returns a slice of each available node of all subconfigurations. Sorted by node id.
+//
+// NOTE: mutating the returned slice is not supported.
+func (c *OnewayTestConfiguration) AllNodes() []*OnewayTestNode {
+	rawNodes := c.RawConfiguration.AllNodes()
+	nodes := make([]*OnewayTestNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &OnewayTestNode{n}
+	}
+	return nodes
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
@@ -87,53 +128,11 @@ func (c OnewayTestConfiguration) Except(rm *OnewayTestConfiguration) gorums.Node
 	return c.RawConfiguration.Except(rm.RawConfiguration)
 }
 
-// OnewayTestManager maintains a connection pool of nodes on
-// which quorum calls can be performed.
-type OnewayTestManager struct {
-	*gorums.RawManager
-}
+// Multicast is a quorum call invoked on each node in configuration c,
 
-// NewOnewayTestManager returns a new OnewayTestManager for managing connection to nodes added
-// to the manager. This function accepts manager options used to configure
-// various aspects of the manager.
-func NewOnewayTestManager(opts ...gorums.ManagerOption) *OnewayTestManager {
-	return &OnewayTestManager{
-		RawManager: gorums.NewRawManager(opts...),
-	}
-}
-
-// NewOnewayTestConfiguration returns a OnewayTestConfiguration based on the provided list of nodes (required).
-// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
-// A new configuration can also be created from an existing configuration,
-// using the And, WithNewNodes, Except, and WithoutNodes methods.
-func (m *OnewayTestManager) NewConfiguration(cfg gorums.NodeListOption) (c *OnewayTestConfiguration, err error) {
-	c = &OnewayTestConfiguration{}
-	c.RawConfiguration, err = gorums.NewRawConfiguration(m.RawManager, cfg)
-	if err != nil {
-		return nil, err
-	}
-	// initialize the nodes slice
-	c.nodes = make([]*OnewayTestNode, c.Size())
-	for i, n := range c.RawConfiguration {
-		c.nodes[i] = &OnewayTestNode{n}
-	}
-	return c, nil
-}
-
-// Nodes returns a slice of available nodes on this manager.
-// IDs are returned in the order they were added at creation of the manager.
-func (m *OnewayTestManager) Nodes() []*OnewayTestNode {
-	gorumsNodes := m.RawManager.Nodes()
-	nodes := make([]*OnewayTestNode, len(gorumsNodes))
-	for i, n := range gorumsNodes {
-		nodes[i] = &OnewayTestNode{n}
-	}
-	return nodes
-}
-
-// Multicast is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
+// with the same argument in, and returns the responses as an iterator.
 func (c *OnewayTestConfiguration) Multicast(ctx context.Context, in *Request, opts ...gorums.CallOption) {
+
 	cd := gorums.QuorumCallData{
 		Message: in,
 		Method:  "oneway.OnewayTest.Multicast",
@@ -143,17 +142,20 @@ func (c *OnewayTestConfiguration) Multicast(ctx context.Context, in *Request, op
 }
 
 // MulticastPerNode is a quorum call invoked on each node in configuration c,
-// with the argument returned by the provided function f, and returns the combined result.
+
+// with the argument returned by the provided function f,
+// it returns the responses as an iterator.
 // The per node function f receives a copy of the Request request argument and
 // returns a Request manipulated to be passed to the given nodeID.
 // The function f must be thread-safe.
 func (c *OnewayTestConfiguration) MulticastPerNode(ctx context.Context, in *Request, f func(*Request, uint32) *Request, opts ...gorums.CallOption) {
+
 	cd := gorums.QuorumCallData{
 		Message: in,
 		Method:  "oneway.OnewayTest.MulticastPerNode",
 	}
 
-	cd.PerNodeArgFn = func(req protoreflect.ProtoMessage, nid uint32) protoreflect.ProtoMessage {
+	cd.PerNodeArgFn = func(req proto.Message, nid uint32) proto.Message {
 		return f(req.(*Request), nid)
 	}
 
