@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
+// 	protoc-gen-gorums v0.9.0-devel+d97bfa83
 // 	protoc            v6.30.2
 // source: qf/qf.proto
 
@@ -8,17 +8,15 @@ package qf
 
 import (
 	context "context"
-	fmt "fmt"
 	gorums "github.com/relab/gorums"
 	encoding "google.golang.org/grpc/encoding"
-	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 func init() {
@@ -29,8 +27,8 @@ func init() {
 
 // QuorumFunctionClient is the client interface for the QuorumFunction service.
 type QuorumFunctionClient interface {
-	UseReq(ctx context.Context, in *Request) (resp *Response, err error)
-	IgnoreReq(ctx context.Context, in *Request) (resp *Response, err error)
+	UseReq(ctx context.Context, in *Request) gorums.Responses[*Response]
+	IgnoreReq(ctx context.Context, in *Request) gorums.Responses[*Response]
 }
 
 // enforce interface compliance
@@ -47,29 +45,47 @@ var _ QuorumFunctionNodeClient = (*QuorumFunctionNode)(nil)
 // procedure calls may be invoked.
 type QuorumFunctionConfiguration struct {
 	gorums.RawConfiguration
-	qspec QuorumFunctionQuorumSpec
-	nodes []*QuorumFunctionNode
 }
 
-// QuorumFunctionQuorumSpecFromRaw returns a new QuorumFunctionQuorumSpec from the given raw configuration and QuorumSpec.
+// NewQuorumFunctionConfiguration returns a configuration based on the provided list of nodes (required)
+// and an optional quorum specification. The QuorumSpec is necessary for call types that
+// must process replies. For configurations only used for unicast or multicast call types,
+// a QuorumSpec is not needed.
+// Nodes can be supplied using WithNodeMap or WithNodeList.
+// Using any other type of NodeListOption will not work.
+// The ManagerOption list controls how the nodes in the configuration are created.
+func NewQuorumFunctionConfiguration(cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *QuorumFunctionConfiguration, err error) {
+	c = &QuorumFunctionConfiguration{}
+	c.RawConfiguration, err = gorums.NewRawConfiguration(cfg, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return c, nil
+}
+
+// SubQuorumFunctionConfiguration allows for making a new Configuration from the
+// ManagerOption list and node list of another set of configurations,
+// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
+// A new configuration can also be created from an existing configuration,
+// using the And, WithNewNodes, Except, and WithoutNodes methods.
+func (c *QuorumFunctionConfiguration) SubQuorumFunctionConfiguration(cfg gorums.NodeListOption) (subCfg *QuorumFunctionConfiguration, err error) {
+	subCfg = &QuorumFunctionConfiguration{}
+	subCfg.RawConfiguration, err = c.SubRawConfiguration(cfg)
+	if err != nil {
+		return nil, err
+	}
+	return subCfg, nil
+}
+
+// QuorumFunctionConfigurationFromRaw returns a new QuorumFunctionConfiguration from the given raw configuration.
 //
 // This function may for example be used to "clone" a configuration but install a different QuorumSpec:
 //
 //	cfg1, err := mgr.NewConfiguration(qspec1, opts...)
 //	cfg2 := ConfigurationFromRaw(cfg1.RawConfig, qspec2)
-func QuorumFunctionConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec QuorumFunctionQuorumSpec) (*QuorumFunctionConfiguration, error) {
-	// return an error if qspec is nil.
-	if qspec == nil {
-		return nil, fmt.Errorf("config: missing required QuorumSpec")
-	}
+func QuorumFunctionConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*QuorumFunctionConfiguration, error) {
 	newCfg := &QuorumFunctionConfiguration{
 		RawConfiguration: rawCfg,
-		qspec:            qspec,
-	}
-	// initialize the nodes slice
-	newCfg.nodes = make([]*QuorumFunctionNode, newCfg.Size())
-	for i, n := range rawCfg {
-		newCfg.nodes[i] = &QuorumFunctionNode{n}
 	}
 	return newCfg, nil
 }
@@ -79,7 +95,24 @@ func QuorumFunctionConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec Qu
 //
 // NOTE: mutating the returned slice is not supported.
 func (c *QuorumFunctionConfiguration) Nodes() []*QuorumFunctionNode {
-	return c.nodes
+	rawNodes := c.RawConfiguration.Nodes()
+	nodes := make([]*QuorumFunctionNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &QuorumFunctionNode{n}
+	}
+	return nodes
+}
+
+// AllNodes returns a slice of each available node of all subconfigurations. Sorted by node id.
+//
+// NOTE: mutating the returned slice is not supported.
+func (c *QuorumFunctionConfiguration) AllNodes() []*QuorumFunctionNode {
+	rawNodes := c.RawConfiguration.AllNodes()
+	nodes := make([]*QuorumFunctionNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &QuorumFunctionNode{n}
+	}
+	return nodes
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
@@ -93,121 +126,35 @@ func (c QuorumFunctionConfiguration) Except(rm *QuorumFunctionConfiguration) gor
 	return c.RawConfiguration.Except(rm.RawConfiguration)
 }
 
-// QuorumFunctionManager maintains a connection pool of nodes on
-// which quorum calls can be performed.
-type QuorumFunctionManager struct {
-	*gorums.RawManager
-}
-
-// NewQuorumFunctionManager returns a new QuorumFunctionManager for managing connection to nodes added
-// to the manager. This function accepts manager options used to configure
-// various aspects of the manager.
-func NewQuorumFunctionManager(opts ...gorums.ManagerOption) *QuorumFunctionManager {
-	return &QuorumFunctionManager{
-		RawManager: gorums.NewRawManager(opts...),
-	}
-}
-
-// NewQuorumFunctionConfiguration returns a QuorumFunctionConfiguration based on the provided list of nodes (required)and a quorum specification.
-// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
-// A new configuration can also be created from an existing configuration,
-// using the And, WithNewNodes, Except, and WithoutNodes methods.
-func (m *QuorumFunctionManager) NewConfiguration(cfg gorums.NodeListOption, qspec QuorumFunctionQuorumSpec) (c *QuorumFunctionConfiguration, err error) {
-	c = &QuorumFunctionConfiguration{}
-	c.RawConfiguration, err = gorums.NewRawConfiguration(m.RawManager, cfg)
-	if err != nil {
-		return nil, err
-	}
-	// return an error if qspec is nil.
-	if qspec == nil {
-		return nil, fmt.Errorf("config: missing required QuorumFunctionQuorumSpec")
-	}
-	c.qspec = qspec
-	// initialize the nodes slice
-	c.nodes = make([]*QuorumFunctionNode, c.Size())
-	for i, n := range c.RawConfiguration {
-		c.nodes[i] = &QuorumFunctionNode{n}
-	}
-	return c, nil
-}
-
-// Nodes returns a slice of available nodes on this manager.
-// IDs are returned in the order they were added at creation of the manager.
-func (m *QuorumFunctionManager) Nodes() []*QuorumFunctionNode {
-	gorumsNodes := m.RawManager.Nodes()
-	nodes := make([]*QuorumFunctionNode, len(gorumsNodes))
-	for i, n := range gorumsNodes {
-		nodes[i] = &QuorumFunctionNode{n}
-	}
-	return nodes
-}
-
 // QuorumFunctionNode holds the node specific methods for the QuorumFunction service.
 type QuorumFunctionNode struct {
 	*gorums.RawNode
 }
 
-// QuorumFunctionQuorumSpec is the interface of quorum functions for QuorumFunction.
-type QuorumFunctionQuorumSpec interface {
-	gorums.ConfigOption
+// UseReq is a quorum call invoked on each node in configuration c,
 
-	// UseReqQF is the quorum function for the UseReq
-	// quorum call method. The in parameter is the request object
-	// supplied to the UseReq method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *Request'.
-	UseReqQF(in *Request, replies map[uint32]*Response) (*Response, bool)
+// with the same argument in, and returns the responses as an iterator.
+func (c *QuorumFunctionConfiguration) UseReq(ctx context.Context, in *Request) gorums.Responses[*Response] {
+	cd := gorums.QuorumCallData{
+		Message:      in,
+		Method:       "qf.QuorumFunction.UseReq",
+		ServerStream: false,
+	}
 
-	// IgnoreReqQF is the quorum function for the IgnoreReq
-	// quorum call method. The in parameter is the request object
-	// supplied to the IgnoreReq method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *Request'.
-	IgnoreReqQF(in *Request, replies map[uint32]*Response) (*Response, bool)
+	return gorums.QuorumCall[*Response](ctx, c.RawConfiguration, cd)
 }
 
-// UseReq is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
-func (c *QuorumFunctionConfiguration) UseReq(ctx context.Context, in *Request) (resp *Response, err error) {
+// IgnoreReq is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
+func (c *QuorumFunctionConfiguration) IgnoreReq(ctx context.Context, in *Request) gorums.Responses[*Response] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "qf.QuorumFunction.UseReq",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Response, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Response)
-		}
-		return c.qspec.UseReqQF(req.(*Request), r)
+		Message:      in,
+		Method:       "qf.QuorumFunction.IgnoreReq",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*Response), err
-}
-
-// IgnoreReq is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
-func (c *QuorumFunctionConfiguration) IgnoreReq(ctx context.Context, in *Request) (resp *Response, err error) {
-	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "qf.QuorumFunction.IgnoreReq",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Response, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Response)
-		}
-		return c.qspec.IgnoreReqQF(req.(*Request), r)
-	}
-
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*Response), err
+	return gorums.QuorumCall[*Response](ctx, c.RawConfiguration, cd)
 }
 
 // QuorumFunction is the server-side API for the QuorumFunction Service
@@ -229,10 +176,4 @@ func RegisterQuorumFunctionServer(srv *gorums.Server, impl QuorumFunctionServer)
 		resp, err := impl.IgnoreReq(ctx, req)
 		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
 	})
-}
-
-type internalResponse struct {
-	nid   uint32
-	reply *Response
-	err   error
 }
