@@ -10,7 +10,6 @@ import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
-	sync "sync"
 	unsafe "unsafe"
 )
 
@@ -22,7 +21,7 @@ const (
 )
 
 type Request struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
+	state         protoimpl.MessageState `protogen:"opaque.v1"`
 	unknownFields protoimpl.UnknownFields
 	sizeCache     protoimpl.SizeCache
 }
@@ -52,14 +51,21 @@ func (x *Request) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Request.ProtoReflect.Descriptor instead.
-func (*Request) Descriptor() ([]byte, []int) {
-	return file_tls_tls_proto_rawDescGZIP(), []int{0}
+type Request_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+}
+
+func (b0 Request_builder) Build() *Request {
+	m0 := &Request{}
+	b, x := &b0, m0
+	_, _ = b, x
+	return m0
 }
 
 type Response struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
-	OK            bool                   `protobuf:"varint,1,opt,name=OK" json:"OK,omitempty"`
+	state         protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_OK bool                   `protobuf:"varint,1,opt,name=OK"`
 	unknownFields protoimpl.UnknownFields
 	sizeCache     protoimpl.SizeCache
 }
@@ -89,16 +95,29 @@ func (x *Response) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Response.ProtoReflect.Descriptor instead.
-func (*Response) Descriptor() ([]byte, []int) {
-	return file_tls_tls_proto_rawDescGZIP(), []int{1}
-}
-
 func (x *Response) GetOK() bool {
 	if x != nil {
-		return x.OK
+		return x.xxx_hidden_OK
 	}
 	return false
+}
+
+func (x *Response) SetOK(v bool) {
+	x.xxx_hidden_OK = v
+}
+
+type Response_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	OK bool
+}
+
+func (b0 Response_builder) Build() *Response {
+	m0 := &Response{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_OK = b.OK
+	return m0
 }
 
 var File_tls_tls_proto protoreflect.FileDescriptor
@@ -111,18 +130,6 @@ const file_tls_tls_proto_rawDesc = "" +
 	"\x02OK\x18\x01 \x01(\bR\x02OK2/\n" +
 	"\x03TLS\x12(\n" +
 	"\aTestTLS\x12\f.tls.Request\x1a\r.tls.Response\"\x00B(Z!github.com/relab/gorums/tests/tls\x92\x03\x02\b\x02b\beditionsp\xe8\a"
-
-var (
-	file_tls_tls_proto_rawDescOnce sync.Once
-	file_tls_tls_proto_rawDescData []byte
-)
-
-func file_tls_tls_proto_rawDescGZIP() []byte {
-	file_tls_tls_proto_rawDescOnce.Do(func() {
-		file_tls_tls_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_tls_tls_proto_rawDesc), len(file_tls_tls_proto_rawDesc)))
-	})
-	return file_tls_tls_proto_rawDescData
-}
 
 var file_tls_tls_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
 var file_tls_tls_proto_goTypes = []any{
