@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
+// 	protoc-gen-gorums v0.9.0-devel+d97bfa83
 // 	protoc            v6.30.2
 // source: tls/tls.proto
 
@@ -14,9 +14,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 func init() {
@@ -29,10 +29,39 @@ func init() {
 // procedure calls may be invoked.
 type TLSConfiguration struct {
 	gorums.RawConfiguration
-	nodes []*TLSNode
 }
 
-// TLSQuorumSpecFromRaw returns a new TLSQuorumSpec from the given raw configuration.
+// NewTLSConfiguration returns a configuration based on the provided list of nodes (required)
+// and an optional quorum specification. The QuorumSpec is necessary for call types that
+// must process replies. For configurations only used for unicast or multicast call types,
+// a QuorumSpec is not needed.
+// Nodes can be supplied using WithNodeMap or WithNodeList.
+// Using any other type of NodeListOption will not work.
+// The ManagerOption list controls how the nodes in the configuration are created.
+func NewTLSConfiguration(cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *TLSConfiguration, err error) {
+	c = &TLSConfiguration{}
+	c.RawConfiguration, err = gorums.NewRawConfiguration(cfg, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return c, nil
+}
+
+// SubTLSConfiguration allows for making a new Configuration from the
+// ManagerOption list and node list of another set of configurations,
+// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
+// A new configuration can also be created from an existing configuration,
+// using the And, WithNewNodes, Except, and WithoutNodes methods.
+func (c *TLSConfiguration) SubTLSConfiguration(cfg gorums.NodeListOption) (subCfg *TLSConfiguration, err error) {
+	subCfg = &TLSConfiguration{}
+	subCfg.RawConfiguration, err = c.SubRawConfiguration(cfg)
+	if err != nil {
+		return nil, err
+	}
+	return subCfg, nil
+}
+
+// TLSConfigurationFromRaw returns a new TLSConfiguration from the given raw configuration.
 //
 // This function may for example be used to "clone" a configuration but install a different QuorumSpec:
 //
@@ -42,11 +71,6 @@ func TLSConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*TLSConfiguration,
 	newCfg := &TLSConfiguration{
 		RawConfiguration: rawCfg,
 	}
-	// initialize the nodes slice
-	newCfg.nodes = make([]*TLSNode, newCfg.Size())
-	for i, n := range rawCfg {
-		newCfg.nodes[i] = &TLSNode{n}
-	}
 	return newCfg, nil
 }
 
@@ -55,7 +79,24 @@ func TLSConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*TLSConfiguration,
 //
 // NOTE: mutating the returned slice is not supported.
 func (c *TLSConfiguration) Nodes() []*TLSNode {
-	return c.nodes
+	rawNodes := c.RawConfiguration.Nodes()
+	nodes := make([]*TLSNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &TLSNode{n}
+	}
+	return nodes
+}
+
+// AllNodes returns a slice of each available node of all subconfigurations. Sorted by node id.
+//
+// NOTE: mutating the returned slice is not supported.
+func (c *TLSConfiguration) AllNodes() []*TLSNode {
+	rawNodes := c.RawConfiguration.AllNodes()
+	nodes := make([]*TLSNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &TLSNode{n}
+	}
+	return nodes
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
@@ -69,57 +110,14 @@ func (c TLSConfiguration) Except(rm *TLSConfiguration) gorums.NodeListOption {
 	return c.RawConfiguration.Except(rm.RawConfiguration)
 }
 
-// TLSManager maintains a connection pool of nodes on
-// which quorum calls can be performed.
-type TLSManager struct {
-	*gorums.RawManager
-}
-
-// NewTLSManager returns a new TLSManager for managing connection to nodes added
-// to the manager. This function accepts manager options used to configure
-// various aspects of the manager.
-func NewTLSManager(opts ...gorums.ManagerOption) *TLSManager {
-	return &TLSManager{
-		RawManager: gorums.NewRawManager(opts...),
-	}
-}
-
-// NewTLSConfiguration returns a TLSConfiguration based on the provided list of nodes (required).
-// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
-// A new configuration can also be created from an existing configuration,
-// using the And, WithNewNodes, Except, and WithoutNodes methods.
-func (m *TLSManager) NewConfiguration(cfg gorums.NodeListOption) (c *TLSConfiguration, err error) {
-	c = &TLSConfiguration{}
-	c.RawConfiguration, err = gorums.NewRawConfiguration(m.RawManager, cfg)
-	if err != nil {
-		return nil, err
-	}
-	// initialize the nodes slice
-	c.nodes = make([]*TLSNode, c.Size())
-	for i, n := range c.RawConfiguration {
-		c.nodes[i] = &TLSNode{n}
-	}
-	return c, nil
-}
-
-// Nodes returns a slice of available nodes on this manager.
-// IDs are returned in the order they were added at creation of the manager.
-func (m *TLSManager) Nodes() []*TLSNode {
-	gorumsNodes := m.RawManager.Nodes()
-	nodes := make([]*TLSNode, len(gorumsNodes))
-	for i, n := range gorumsNodes {
-		nodes[i] = &TLSNode{n}
-	}
-	return nodes
-}
-
 // TLSNode holds the node specific methods for the TLS service.
 type TLSNode struct {
 	*gorums.RawNode
 }
 
-// TestTLS is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
+// TestTLS is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
 func (n *TLSNode) TestTLS(ctx context.Context, in *Request) (resp *Response, err error) {
 	cd := gorums.CallData{
 		Message: in,
