@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
+// 	protoc-gen-gorums v0.9.0-devel+d97bfa83
 // 	protoc            v6.30.2
 // source: unresponsive/unresponsive.proto
 
@@ -14,9 +14,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 func init() {
@@ -29,10 +29,39 @@ func init() {
 // procedure calls may be invoked.
 type UnresponsiveConfiguration struct {
 	gorums.RawConfiguration
-	nodes []*UnresponsiveNode
 }
 
-// UnresponsiveQuorumSpecFromRaw returns a new UnresponsiveQuorumSpec from the given raw configuration.
+// NewUnresponsiveConfiguration returns a configuration based on the provided list of nodes (required)
+// and an optional quorum specification. The QuorumSpec is necessary for call types that
+// must process replies. For configurations only used for unicast or multicast call types,
+// a QuorumSpec is not needed.
+// Nodes can be supplied using WithNodeMap or WithNodeList.
+// Using any other type of NodeListOption will not work.
+// The ManagerOption list controls how the nodes in the configuration are created.
+func NewUnresponsiveConfiguration(cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *UnresponsiveConfiguration, err error) {
+	c = &UnresponsiveConfiguration{}
+	c.RawConfiguration, err = gorums.NewRawConfiguration(cfg, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return c, nil
+}
+
+// SubUnresponsiveConfiguration allows for making a new Configuration from the
+// ManagerOption list and node list of another set of configurations,
+// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
+// A new configuration can also be created from an existing configuration,
+// using the And, WithNewNodes, Except, and WithoutNodes methods.
+func (c *UnresponsiveConfiguration) SubUnresponsiveConfiguration(cfg gorums.NodeListOption) (subCfg *UnresponsiveConfiguration, err error) {
+	subCfg = &UnresponsiveConfiguration{}
+	subCfg.RawConfiguration, err = c.SubRawConfiguration(cfg)
+	if err != nil {
+		return nil, err
+	}
+	return subCfg, nil
+}
+
+// UnresponsiveConfigurationFromRaw returns a new UnresponsiveConfiguration from the given raw configuration.
 //
 // This function may for example be used to "clone" a configuration but install a different QuorumSpec:
 //
@@ -42,11 +71,6 @@ func UnresponsiveConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*Unrespon
 	newCfg := &UnresponsiveConfiguration{
 		RawConfiguration: rawCfg,
 	}
-	// initialize the nodes slice
-	newCfg.nodes = make([]*UnresponsiveNode, newCfg.Size())
-	for i, n := range rawCfg {
-		newCfg.nodes[i] = &UnresponsiveNode{n}
-	}
 	return newCfg, nil
 }
 
@@ -55,7 +79,24 @@ func UnresponsiveConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*Unrespon
 //
 // NOTE: mutating the returned slice is not supported.
 func (c *UnresponsiveConfiguration) Nodes() []*UnresponsiveNode {
-	return c.nodes
+	rawNodes := c.RawConfiguration.Nodes()
+	nodes := make([]*UnresponsiveNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &UnresponsiveNode{n}
+	}
+	return nodes
+}
+
+// AllNodes returns a slice of each available node of all subconfigurations. Sorted by node id.
+//
+// NOTE: mutating the returned slice is not supported.
+func (c *UnresponsiveConfiguration) AllNodes() []*UnresponsiveNode {
+	rawNodes := c.RawConfiguration.AllNodes()
+	nodes := make([]*UnresponsiveNode, len(rawNodes))
+	for i, n := range rawNodes {
+		nodes[i] = &UnresponsiveNode{n}
+	}
+	return nodes
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
@@ -69,57 +110,14 @@ func (c UnresponsiveConfiguration) Except(rm *UnresponsiveConfiguration) gorums.
 	return c.RawConfiguration.Except(rm.RawConfiguration)
 }
 
-// UnresponsiveManager maintains a connection pool of nodes on
-// which quorum calls can be performed.
-type UnresponsiveManager struct {
-	*gorums.RawManager
-}
-
-// NewUnresponsiveManager returns a new UnresponsiveManager for managing connection to nodes added
-// to the manager. This function accepts manager options used to configure
-// various aspects of the manager.
-func NewUnresponsiveManager(opts ...gorums.ManagerOption) *UnresponsiveManager {
-	return &UnresponsiveManager{
-		RawManager: gorums.NewRawManager(opts...),
-	}
-}
-
-// NewUnresponsiveConfiguration returns a UnresponsiveConfiguration based on the provided list of nodes (required).
-// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
-// A new configuration can also be created from an existing configuration,
-// using the And, WithNewNodes, Except, and WithoutNodes methods.
-func (m *UnresponsiveManager) NewConfiguration(cfg gorums.NodeListOption) (c *UnresponsiveConfiguration, err error) {
-	c = &UnresponsiveConfiguration{}
-	c.RawConfiguration, err = gorums.NewRawConfiguration(m.RawManager, cfg)
-	if err != nil {
-		return nil, err
-	}
-	// initialize the nodes slice
-	c.nodes = make([]*UnresponsiveNode, c.Size())
-	for i, n := range c.RawConfiguration {
-		c.nodes[i] = &UnresponsiveNode{n}
-	}
-	return c, nil
-}
-
-// Nodes returns a slice of available nodes on this manager.
-// IDs are returned in the order they were added at creation of the manager.
-func (m *UnresponsiveManager) Nodes() []*UnresponsiveNode {
-	gorumsNodes := m.RawManager.Nodes()
-	nodes := make([]*UnresponsiveNode, len(gorumsNodes))
-	for i, n := range gorumsNodes {
-		nodes[i] = &UnresponsiveNode{n}
-	}
-	return nodes
-}
-
 // UnresponsiveNode holds the node specific methods for the Unresponsive service.
 type UnresponsiveNode struct {
 	*gorums.RawNode
 }
 
-// TestUnresponsive is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
+// TestUnresponsive is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
 func (n *UnresponsiveNode) TestUnresponsive(ctx context.Context, in *Empty) (resp *Empty, err error) {
 	cd := gorums.CallData{
 		Message: in,
